Struct bitvec::array::BitArray

#[repr(transparent)]pub struct BitArray<O = Lsb0, V = [usize; 1]> where
    O: BitOrder,
    V: BitView,  { /* fields omitted */ }

An array of individual bits, able to be held by value on the stack.

This type is generic over all Sized implementors of the BitView trait. Due to limitations in the Rust language’s const-generics implementation (it is both unstable and incomplete), this must take an array type parameter directly, rather than register type and bit-count integer parameters. This makes it less convenient to use than C++’s std::bitset<N> array type. The bitarr! macro is capable of constructing both values and specific types of BitArray, and this macro should be preferred for most use.

The advantage of using this wrapper is that it implements Deref/Mut to BitSlice, as well as implementing all of BitSlices traits by forwarding to the BitSlice view of its contained data. This allows it to have BitSlice behavior by itself, without requiring explicit .as_bitslice() calls in user code.

Limitations

This does not track start or end indices of its BitSlice view, and so that view will always fully span the buffer. You cannot produce, for example, an array of twelve bits.

Type Parameters

O: The ordering of bits within memory registers.

V: Some buffer which can be used as the basis for a BitSlice view. This will usually be an array of [T: BitRegister; N].

Examples

This type is useful for marking that some value is always to be used as a BitSlice.

use bitvec::prelude::*;

struct HasBitfields {
  header: u32,
  // creates a type declaration.
  fields: bitarr!(for 20, in Msb0, u8),
}

impl HasBitfields {
  pub fn new() -> Self {
    Self {
      header: 0,
      // creates a value object.
      // the type paramaters must be repeated.
      fields: bitarr![Msb0, u8; 0; 20],
    }
  }

  /// Access a bit region directly
  pub fn get_subfield(&self) -> &BitSlice<Msb0, u8> {
    &self.fields[.. 4]
  }

  /// Read a 12-bit value out of a region
  pub fn read_value(&self) -> u16 {
    self.fields[4 .. 16].load()
  }

  /// Write a 12-bit value into a region
  pub fn set_value(&mut self, value: u16) {
    self.fields[4 .. 16].store(value);
  }
}

Eventual Obsolescence

When const-generics stabilize, this will be modified to have a signature more like BitArray<O, T, const N: usize>([T; elts::<T>(N)]);, to mirror the behavior of ordinary arrays [T; N] as they stand today.

Implementations

impl<O, V> BitArray<O, V> where
O: BitOrder,
V: BitView,

pub fn zeroed() -> Self

Constructs a new BitArray with its memory set to zero.

pub fn new(data: V) -> Self

Wraps a buffer in a BitArray.

Examples

use bitvec::prelude::*;

let data = [0u8; 2];
let bits: BitArray<Msb0, _> = BitArray::new(data);
assert_eq!(bits.len(), 16);

pub fn value(self) -> V

Removes the BitArray wrapper, leaving the contained buffer.

Examples

use bitvec::prelude::*;

let bitarr = bitarr![Lsb0, usize; 0; 30];
let native: [usize; 1] = bitarr.value();

pub fn as_bitslice(&self) -> &BitSlice<O, V::Store>

Views the array as a BitSlice.

pub fn as_mut_bitslice(&mut self) -> &mut BitSlice<O, V::Store>

Views the array as a mutable BitSlice.

pub fn as_raw_slice(&self) -> &[V::Store]

Views the array as a slice of its underlying memory registers.

pub fn as_mut_raw_slice(&mut self) -> &mut [V::Store]

Views the array as a mutable slice of its underlying memory registers.

pub fn as_buffer(&self) -> &V

Views the interior buffer.

pub fn as_mut_buffer(&mut self) -> &mut V

Mutably views the interior buffer.