Trait bitvec::bits::BitsMut

pub trait BitsMut: Bits {
    fn as_mut_bitslice<C>(&mut self) -> &mut BitSlice<C, Self::Store>
    where
        C: Cursor;
}

Allows a type to be used as a sequence of mutable bits.

Requirements

This trait can only be implemented by contiguous structures: individual fundamentals, and sequences (arrays or slices) of them.

Required methods

fn as_mut_bitslice<C>(&mut self) -> &mut BitSlice<C, Self::Store> where
C: Cursor,

Constructs a mutable BitSlice reference over data.

Type Parameters

C: Cursor: The Cursor type used to index within the slice.

Parameters

&mut self

Returns

A BitSlice handle over self’s data, using the provided Cursor type and using Self::Store as the data type.

Examples

use bitvec::prelude::*;

let mut src = 8u8;
let bits = src.as_mut_bitslice::<LittleEndian>();
assert!(bits[3]);
*bits.at(3) = false;
assert!(!bits[3]);

Loading content...