bitvec::boxed
Struct BitBox

pub struct BitBox<O = Lsb0, T = usize>
where
    O: BitOrder,
    T: BitStore,
{ /* private fields */ }

A frozen heap-allocated buffer of individual bits.

This is essentially a BitVec that has frozen its allocation, and given up the ability to change size. It is analagous to Box<[bool]>. You should prefer BitArray over BitBox where possible, and may freely box it if you need the indirection.

Documentation

All APIs that mirror something in the standard library will have an Original section linking to the corresponding item. All APIs that have a different signature or behavior than the original will have an API Differences section explaining what has changed, and how to adapt your existing code to the change.

These sections look like this:

Original

Box<[T]>

API Differences

The buffer type Box<[bool]> has no type parameters. BitBox<O, T> has the same two type parameters as BitSlice<O, T>. Otherwise, BitBox is able to implement the full API surface of Box<[bool]>.

Behavior

Because BitBox is a fully-owned buffer, it is able to operate on its memory without concern for any other views that may alias. This enables it to specialize some BitSlice behavior to be faster or more efficient.

Type Parameters

This takes the same BitOrder and BitStore parameters as BitSlice. Unlike BitSlice, it is restricted to only accept the fundamental integers as its BitStore arguments; BitBox buffers can never be aliased by other BitBoxes, and do not need to share memory access.

Safety

BitBox is a wrapper over a NonNull<BitSlice<O, T>> pointer; this allows it to remain exactly two words in size, and means that it is subject to the same representational incompatibility restrictions as BitSlice references. You must never attempt to type-cast between Box<[bool]> and BitBox in any way, nor may you attempt to modify the memory value of a BitBox handle. Doing so will cause allocator and memory errors in your program, likely inducing a panic.

Everything in the BitBox public API, even the unsafe parts, are guaranteed to have no more unsafety or potential for incorrectness than their equivalent items in the standard library. All unsafe APIs will have documentation explicitly detailing what the API requires you to uphold in order for it to function safely and correctly. All safe APIs will do so themselves.

Macro Construction

Heap allocation can only occur at runtime, but the bitbox! macro will construct an appropriate BitSlice buffer at compile-time, and at run-time, only copy the buffer into a heap allocation.

Implementations

impl<O, T> BitBox<O, T>
where O: BitOrder, T: BitStore,

pub fn new(x: &BitSlice<O, T>) -> Self

👎Deprecated: Prefer from_bitslice

Allocates memory on the heap and then copies x into it.

This doesn’t actually allocate if x is zero-length.

Original

Box::new

API Differences

Box::<[T]>::new does not exist, because unsized types cannot be taken by value. Instead, this takes a slice reference, and boxes the referent slice.

Examples

use bitvec::prelude::*;

let boxed = BitBox::new(bits![0; 5]);

pub fn pin(x: &BitSlice<O, T>) -> Pin<Self>
where O: Unpin, T: Unpin,

Constructs a new Pin<BitBox<O, T>>.

BitSlice is always Unpin, so this has no actual effect.

Original

Box::pin

API Differences

As with new, this only exists on Box when T is not unsized. This takes a slice reference, and pins the referent slice.

pub unsafe fn from_raw(raw: *mut BitSlice<O, T>) -> Self

Constructs a bit-box from a raw bit-slice pointer.

After calling this function, the raw bit-slice pointer is owned by the resulting BitBox. Specifically, the BitBox destructor will free the memory allocation at the bit-slice pointer’s address. For this to be safe, the bit-slice pointer can only have been produced by a BitBox previously destroyed using into_raw.

Original

Box::from_raw

Safety

This function is unsafe because improper use may lead to memory problems. For example, a double-free may occur if the function is called twice on the same raw bit-slice pointer.

Examples

Recreate a BitBox which was previously converted to a raw bit-slice pointer using BitBox::into_raw:

use bitvec::prelude::*;

let x = bitbox![0; 10];
let ptr = BitBox::into_raw(x);
let x = unsafe { BitBox::from_raw(ptr) };

pub fn into_raw(this: Self) -> *mut BitSlice<O, T>

Consumes the BitBox, returning a raw bit-slice pointer.

The bit-slice pointer will be properly encoded and non-null.

After calling this function, the caller is responsible for the memory previously managed by the BitBox. In particular, the caller should properly release the memory by converting the bit-slice pointer back into a BitBox with the from_raw function, allowing the BitBox destructor to perform the cleanup.

Note: this is an associated function, which means that you have to call it as BitBox::into_raw(b) instead of b.into_raw(). This is to match signatures with the standard library’s Box API; there will never be a name conflict with BitSlice.

Original

Box::into_raw

Examples

Converting the raw bit-slice pointer back into a BitBox with BitBox::from_raw for automatic cleanup:

use bitvec::prelude::*;

let x = bitbox![0; 50];
let p = BitBox::into_raw(x);
let x = unsafe { BitBox::from_raw(p) };

You may not deällocate pointers produced by this function through any other manner.

pub fn leak<'a>(this: Self) -> &'a mut BitSlice<O, T>
where T: 'a,

Consumes and leaks the BitBox, returning a mutable reference, &'a mut BitSlice<O, T>. This is eligible to be promoted to the 'static lifetime.

Original

Box::leak

This function is mainly useful for data that lives for the remainder of the program’s life. Dropping the returned reference will cause a memory leak. If this is not acceptable, the reference should first be wrapped with the BitBox::from_raw function producing a BitBox. This BitBox can then be dropped which will properly deällocate the memory.

Note: this is an associated function, which means that you have to call it as BitBox::leak(b) instead of b.leak(). This is to match signatures with the standard library’s Box API; there will never be a name conflict with BitSlice.

Original

Box::leak

Examples

Simple usage:

use bitvec::prelude::*;

let b = bitbox![0; 50];
let static_ref: &'static mut BitSlice = BitBox::leak(b);
static_ref.set(0, true);
assert!(static_ref[0]);

impl<O, T> BitBox<O, T>
where O: BitOrder, T: BitStore,

General-purpose functions not present on Box<[T]>.

pub fn from_bitslice(slice: &BitSlice<O, T>) -> Self

Copies a BitSlice region into a new BitBox allocation.

Effects

This delegates to BitVec::from_bitslice, then discards the excess capacity.

Examples

use bitvec::prelude::*;

let bits = bits![0, 1, 0, 1, 1, 0, 1, 1];
let bb = BitBox::from_bitslice(&bits[2 ..]);
assert_eq!(bb, bits[2 ..]);
assert_eq!(bb.as_slice(), bits.as_slice());

pub fn from_boxed_slice(boxed: Box<[T]>) -> Self

Converts a Box<[T]> into a BitBox<O, T>` without copying its buffer.

Parameters

boxed: A boxed slice to view as bits.

Returns

A BitBox over the boxed buffer.

Panics

This panics if boxed is too long to convert into a BitBox. See BitSlice::MAX_ELTS.

Examples

use bitvec::prelude::*;

let boxed: Box<[u8]> = Box::new([0; 4]);
let addr = boxed.as_ptr();
let bb = BitBox::<LocalBits, _>::from_boxed_slice(boxed);
assert_eq!(bb, bits![0; 32]);
assert_eq!(addr, bb.as_slice().as_ptr());

pub fn try_from_boxed_slice(boxed: Box<[T]>) -> Result<Self, Box<[T]>>

Converts a Box<[T]> into a BitBox<O, T> without copying its buffer.

This method takes ownership of a memory buffer and enables it to be used as a bit-box. Because Box<[T]> can be longer than BitBoxes, this is a fallible method, and the original box will be returned if it cannot be converted.

Parameters

boxed: Some boxed slice of memory, to be viewed as bits.

Returns

If boxed is short enough to be viewed as a BitBox, then this returns a BitBox over the boxed buffer. If boxed is too long, then this returns boxed unmodified.

Examples

use bitvec::prelude::*;

let boxed: Box<[u8]> = Box::new([0; 4]);
let addr = boxed.as_ptr();
let bb = BitBox::<LocalBits, _>::try_from_boxed_slice(boxed).unwrap();
assert_eq!(bb[..], bits![0; 32]);
assert_eq!(addr, bb.as_slice().as_ptr());

pub fn into_boxed_slice(self) -> Box<[T]>

Converts the slice back into an ordinary slice of memory elements.

This does not affect the slice’s buffer, only the handle used to control it.

Parameters

self

Returns

An ordinary boxed slice containing all of the bit-slice’s memory buffer.

Examples

use bitvec::prelude::*;

let bb = bitbox![0; 5];
let addr = bb.as_slice().as_ptr();
let boxed = bb.into_boxed_slice();
assert_eq!(boxed[..], [0][..]);
assert_eq!(addr, boxed.as_ptr());

pub fn into_bitvec(self) -> BitVec<O, T>

Converts self into a vector without clones or allocation.

The resulting vector can be converted back into a box via BitVec<O, T>’s .into_boxed_bitslice() method.

Original

slice::into_vec

API Differences

Despite taking a Box<[T]> receiver, this function is written in an impl<T> [T] block.

Rust does not allow the text

impl<O, T> BitSlice<O, T> {
  fn into_bitvec(self: BitBox<O, T>);
}

to be written, and BitBox exists specifically because Box<BitSlice<>> cannot be written either, so this function must be implemented directly on BitBox rather than on BitSlice with a boxed receiver.

Examples

use bitvec::prelude::*;

let bb = bitbox![0, 1, 0, 1];
let bv = bb.into_bitvec();

assert_eq!(bv, bitvec![0, 1, 0, 1]);

pub fn as_bitslice(&self) -> &BitSlice<O, T>

Views the buffer’s contents as a BitSlice.

This is equivalent to &bb[..].

Examples

use bitvec::prelude::*;

let bb = bitbox![0, 1, 1, 0];
let bits = bb.as_bitslice();

pub fn as_mut_bitslice(&mut self) -> &mut BitSlice<O, T>

Extracts a mutable bit-slice of the entire vector.

Equivalent to &mut bv[..].

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 1];
let bits = bv.as_mut_bitslice();
bits.set(0, true);

pub fn as_slice(&self) -> &[T]

Extracts an element slice containing the entire box.

Analogue

See .as_bitslice() for a &BitBox -> &BitSlice transform.

Examples

use bitvec::prelude::*;
use std::io::{self, Write};
let buffer = bitbox![Msb0, u8; 0, 1, 0, 1, 1, 0, 0, 0];
io::sink().write(buffer.as_slice()).unwrap();

pub fn as_mut_slice(&mut self) -> &mut [T]

Extracts a mutable slice of the entire box.

Analogue

See .as_mut_bitslice() for a &mut BitBox -> &mut BitSlice transform.

Examples

use bitvec::prelude::*;
use std::io::{self, Read};
let mut buffer = bitbox![Msb0, u8; 0; 24];
io::repeat(0b101).read_exact(buffer.as_mut_slice()).unwrap();

pub fn set_uninitialized(&mut self, value: bool)

Sets the uninitialized bits of the vector to a fixed value.

This method modifies all bits in the allocated buffer that are outside the self.as_bitslice() view so that they have a consistent value. This can be used to zero the uninitialized memory so that when viewed as a raw memory slice, bits outside the live region have a predictable value.

Examples

use bitvec::prelude::*;

let mut bb = BitBox::new(&220u8.view_bits::<Lsb0>()[.. 4]);
assert_eq!(bb.count_ones(), 2);
assert_eq!(bb.as_slice(), &[220u8]);

bb.set_uninitialized(false);
assert_eq!(bb.as_slice(), &[12u8]);

bb.set_uninitialized(true);
assert_eq!(bb.as_slice(), &[!3u8]);