bitvec
Module domain

Representations of the BitSlice region memory model.

This module allows any BitSlice region to be decomposed into domains that restricts T::Alias markers to only the edge elements that may require them.

Specifically, any given BitSlice region is one of:

touches only interior indices of one element

touches at least one edge index of any number of elements (including zero)

In the latter case, any elements completely spanned by the BitSlice handle are known to not have any other write-capable handles to them, and in the case of an &mut BitSlice handle specifically, no other views at all. As such, the domain view of this memory is able to remove the aliasing marker type and permit direct memory access to the underlying buffer for the duration of its existence.

!

Enums

BitDomain

Granular representation of the memory region containing a BitSlice.

BitDomainMut

Granular representation of the memory region containing a BitSlice.

Domain

Granular representation of the memory region containing a BitSlice.

DomainMut

Granular representation of the memory region containing a BitSlice.