Struct bitvec::index::BitMask

pub struct BitMask<R = usize>where
    R: BitRegister,{ /* private fields */ }

Multi-Bit Selection Mask

Unlike BitSel, which enforces a strict one-hot mask encoding, this type permits any number of bits to be set or cleared. This is used to accumulate selections for batched operations on a register in real memory.

Type Parameters

R: The register element that this mask governs.

Construction

This must only be constructed by combining BitSel selection masks produced through the accepted chains of custody beginning with BitIdx values. Bit-masks not constructed in this manner are not guaranteed to be correct in the caller’s context and may lead to incorrect memory behaviors.

Implementations

impl<R> BitMask<R>where R: BitRegister,

pub const ALL: Self = _

A full bit-mask with every bit set.

pub const ZERO: Self = _

An empty bit-mask with every bit cleared.

pub fn new(mask: R) -> Self

Wraps any R value as a bit-mask.

This constructor is provided to explicitly declare that an operation is discarding the numeric value of an integer and instead using it only as a bit-mask.

Parameters

mask: Some integer to use as a bit-mask.

Returns

The mask value wrapped as a bit-mask, with its numeric context discarded.

Prefer accumulating BitSel values using its Sum implementation.

Safety

The mask value must be computed from a set of valid bit positions in the caller’s context.

pub fn into_inner(self) -> R

Removes the mask wrapper, leaving the internal value.

pub fn test(&self, sel: BitSel<R>) -> bool

Tests if a mask contains a given selector bit.

Parameters

&self: The mask being tested.

sel: A selector bit to test in self.

Returns

Whether self has set the bit that sel indicates.

pub fn insert(&mut self, sel: BitSel<R>)

Inserts a selector bit into a mask.

Parameters

&mut self: The mask being modified.

sel: A selector bit to insert into self.

Effects

The sel bit is set in the mask.

pub fn combine(self, sel: BitSel<R>) -> Self

Creates a new mask with a selector bit activated.

Parameters

self: The original mask.

sel: The selector bit being added into the mask.

Returns

A new bit-mask with sel activated.