Struct bitvec::index::BitSel

#[repr(transparent)]
pub struct BitSel<R> where
    R: BitRegister,  { /* fields omitted */ }

A one-hot selection mask, to be applied to a register R.

This type selects exactly one bit, and is produced by the conversion of a semantic BitIdx to a BitPos through a BitOrder implementation, and then applying 1 << pos. Values of this type are used to select only the bit specified by a BitIdx when performing memory operations.

Type Parameters

R: The register type that these values govern.

Validity

Values of this type are required to have exactly one bit set to 1 and all other bits set to 0.

Construction

This type is only constructed from BitPos values, which are themselves only constructed by a chain of known-good BitIdx values passed into known-correct BitOrder implementations. As such, bitvec can use BitSel values with full confidence that they are correct in the surrounding context.

Implementations

impl<R> BitSel<R> where
R: BitRegister,

pub unsafe fn new(sel: R) -> Option<Self>

Wraps a selector value as a known-good selection of an R register.

Parameters

sel: A one-hot selection mask of a bit in an R register.

Returns

If sel does not have exactly one bit set, this returns None; otherwise, it returns a BitSel wrapping the sel value.

Safety

This function must only be called within a BitOrder::select implementation that is verified to be correct.

pub unsafe fn new_unchecked(sel: R) -> Self

Wraps a selector value as an assumed-good selection of an R register.

Parameters

sel: A one-hot selection mask of a bit in an R register.

Returns

sel wrapped in a BitSel.

Safety

sel must have exactly one bit set high and all others low. In debug builds, invalid sel values cause a panic; release builds do not check the input.

This function must only be called in a correct BitOrder::select implementation.

pub fn mask(self) -> BitMask<R> where
R: BitRegister,

Converts the selector into a bit mask.

This is a type-cast.

pub fn value(self) -> R

Views the internal selector value.

pub fn range_all(
) -> impl Iterator<Item = Self> + DoubleEndedIterator + ExactSizeIterator + FusedIterator

Ranges over all possible selector values.