bitvec::mem
Trait BitMemory

pub trait BitMemory: IsUnsigned + Sealed {
    const BITS: u8 = _;
    const INDX: u8 = _;
    const MASK: u8 = _;
}

Description of an integer memory element.

This trait provides information used to describe integer-typed regions of memory and enables other parts of the project to adequately describe the memory bus. This trait has no bearing on the processor instructions or registers used to interact with memory. It solely describes integers that can exist on a system.

This trait cannot be implemented outside this crate.

Provided Associated Constants

const BITS: u8 = _

The bit width of the integer.

mem::size_of returns the size in bytes, and bytes are always eight bits wide on architectures that Rust targets.

Issue #76904 will place this constant on the fundamental integers directly, as a u32.

const INDX: u8 = _

The number of bits required to store an index in the range 0 .. BITS.

const MASK: u8 = _

A mask over all bits that can be used as an index within the element. This is the value with the least significant INDX-many bits set high.

Object Safety

This trait is not object safe.