Struct bitvec::ptr::BitRef

#[repr(C, align(8))]
pub struct BitRef<'a, M, O = Lsb0, T = usize> where
    M: Mutability,
    O: BitOrder,
    T: BitStore,  { /* private fields */ }

A proxy reference, equivalent to C++ std::bitset<N>::reference.

This type wraps a BitPtr and caches a bool in a padding byte. It is then able to freely produce references to the cached bool, and commits the cache back to the referent bit location on drop.

Lifetimes

'a: The lifetime of the source &'a mut BitSlice that created the BitRef.

Type Parameters

M: The write permission of the reference. When this is Const, the DerefMut implementation is removed, forbidding the proxy from writing back to memory.

O: The ordering used to address a bit in memory.

T: The storage type containing the referent bit.

Quirks

Because this type has both a lifetime and a destructor, it can introduce an uncommon error condition in Rust. When an expression that produces this type is in the final expression of a block, including if that expression is used as a condition in a match, if let, or if, then the compiler will attempt to extend the drop scope of this type to the outside of the block. This causes a lifetime mismatch error if the source region from which this proxy is produced goes out of scope at the end of the block.

If you get a compiler error that this type causes something to be dropped while borrowed, you can end the borrow by putting any expression-ending syntax element after the offending expression that produces this type, including a semicolon or an item definition.

Examples

use bitvec::prelude::*;

let bits = bits![mut 0; 2];

let (left, right) = bits.split_at_mut(1);
let mut first = left.get_mut(0).unwrap();
let second = right.get_mut(0).unwrap();

// Referential behavior
*first = true;
// Direct write
second.set(true);

drop(first); // it’s not a reference!
assert_eq!(bits, bits![1; 2]);

Implementations

impl<M, O, T> BitRef<'_, M, O, T> where
M: Mutability,
O: BitOrder,
T: BitStore,

pub unsafe fn from_bitptr(bitptr: BitPtr<M, O, T>) -> Self

Converts a bit-pointer into a proxy bit-reference.

The conversion reads from the pointer, then stores the bool in a padding byte.

Parameters

bitptr: A bit-pointer to turn into a bit-reference.

Returns

A bit-reference pointing at bitptr.

Safety

The bitptr must address a location that is valid for reads and, if M is Mut, writes.

pub fn into_bitptr(self) -> BitPtr<M, O, T>

Decays the bit-reference to an ordinary bit-pointer.

Parameters

self

Returns

The interior bit-pointer, without the associated cache. If this was a write-capable pointer, then the cached bit is committed to memory before this method returns.

impl<O, T> BitRef<'_, Mut, O, T> where
O: BitOrder,
T: BitStore,

pub fn replace(&mut self, src: bool) -> bool

Moves src into the referenced bit, returning the previous value.

Original

mem::replace

pub fn swap<O2, T2>(&mut self, other: &mut BitRef<'_, Mut, O2, T2>) where
O2: BitOrder,
T2: BitStore,

Swaps the values at two mutable locations, without deïnitializing either one.

Original

mem::swap

pub fn set(self, value: bool)

Writes a bit into the proxied location without an intermediate copy.

This function writes value directly into the proxied location, and does not store value in the proxy’s internal cache. This should be equivalent to the behavior seen when using ordinary DerefMut proxying, but the latter depends on compiler optimization.

Parameters

self: This destroys the proxy, as it becomes invalid when writing directly to the location without updating the cache.

value: The new bit to write into the proxied slot.