Struct bitvec::slice::RSplitMut

pub struct RSplitMut<'a, O, T, P> where
    O: BitOrder,
    T: BitStore,
    P: FnMut(usize, &bool) -> bool,  { /* fields omitted */ }

An iterator over subslices separated by bits that match a predicate function, starting from the end of the BitSlice.

This struct is created by the .rsplit_mut() method on BitSlices.

Original

slice::RSplitMut

API Differences

In order to allow more than one bit of information for the split decision, the predicate receives the index of each bit, as well as its value.

Implementations

impl<'a, O, T, P> RSplitMut<'a, O, T, P> where
O: BitOrder,
T: BitStore,
P: FnMut(usize, &bool) -> bool,

pub unsafe fn remove_alias(self) -> RSplitMutNoAlias<'a, O, T, P>

Adapts the iterator to no longer mark its yielded items as aliased.

Safety

This adapter can only be used in contexts where only one yielded item will be alive at any time. This is most commonly true in for … in loops, so long as no subsequent adapter collects multiple yielded items into a collection where they are live simultaneously.

The items yielded by this iterator will not have an additional alias marker applied to them, so their use in an iteration sequence will not be penalized when the surrounding code guarantees that each item yielded by the iterator is destroyed before the next is produced.

This adapter does not convert the iterator to use T::Mem raw types, as it can be applied to an iterator over an already-aliased slice and must preserve its condition. Its only effect is to prevent the addition of a new T::Alias marker.