Struct bitvec::slice::Split

pub struct Split<'a, O, T, P> where
    O: BitOrder,
    T: 'a + BitStore,
    P: FnMut(usize, &bool) -> bool,  { /* private fields */ }

An iterator over subslices separated by bits that match a predicate function.

This struct is created by the split method on bit slices.

Original

slice::Split

API Differences

In order to allow more than one bit of information for the split decision, the predicate receives the index of each bit, as well as its value.