Trait bitvec::view::BitViewSized

pub trait BitViewSized: BitView + Sized {
    const ELTS: usize;
    const BITS: usize;
}

Helper for size awareness on Sized storage regions.

Associated Constants

const ELTS: usize

Counts the number of elements T contained in the type.

const BITS: usize

Counts the number of bits contained in the type.