Module bitvec::array

A fixed-size region viewed as individual bits, corresponding to [bool].

You can read the language’s array fundamental documentation here.

This module defines the BitArray immediate type, and its associated support code.

BitArray has little behavior or properties in its own right. It serves solely as a type capable of being used in immediate value position, and delegates to BitSlice for all actual work.

Structs

BitArray

An array of individual bits, able to be held by value on the stack.