Module bitvec::boxed

A dynamically-allocated, fixed-size, buffer containing a BitSlice region.

You can read the standard library’s alloc::boxed module documentation here.

This module defines the BitBox buffer, and all of its associated support code.

BitBox is equivalent to Box<[bool]>, in its operation and in its relationship to the BitSlice and BitVec types. Most of the interesting work to be done on a bit-sequence is implemented in BitSlice, to which BitBox dereferences, and the box container itself only exists to maintain wonership and provide some specializations that cannot safely be done on BitSlice alone.

There is almost never a reason to use this type, as it is a mixture of BitArray’s fixed width and BitVec’s heap allocation. You should only use it when you have a bit-sequence whose width is either unknowable at compile-time or inexpressable in BitArray, and are constructing the sequence in a BitVec before freezing it.

!

Structs

BitBox

A frozen heap-allocated buffer of individual bits.