Struct bitvec::boxed::BitBox

#[repr(C)]pub struct BitBox<O = Local, T = usize> where
    O: BitOrder,
    T: BitStore,  { /* fields omitted */ }

A pointer type for owned bit sequences.

This type is essentially a &BitSlice that owns its own memory. It can change the contents of its domain, but it cannot change its own domain like BitVec can. It is useful for fixed-size collections without lifetime tracking.

Type Parameters

O: BitOrder: An implementor of the BitOrder trait. This type is used to convert semantic indices into concrete bit positions in elements, and store or retrieve bit values from the storage type.

T: BitStore: An implementor of the BitStore trait: u8, u16, u32, or u64 (64-bit systems only). This is the actual type in memory that the box will use to store data.

Safety

The BitBox handle has the same size as standard Rust Box<[T]> handles, but it is extremely binary incompatible with them. Attempting to treat BitBox<_, T> as Box<[T]> in any manner except through the provided APIs is catastrophically unsafe and unsound.