Module bitvec::domain

Representation of the BitSlice region memory model

This module allows any BitSlice region to be decomposed into domains with more detailed aliasing information.

Specifically, any particular BitSlice region is one of:

touches only interior indices of one element

touches at least one edge index of any number of elements (including zero)

In the latter case, any elements completely spanned by the slice handle are known to not have any other write-capable views to them, and in the case of an &mut BitSlice handle specifically, no other views at all. As such, the domain view of this memory is able to remove the aliasing marker type and permit direct memory access for the duration of its existence. !

Enums

BitDomain

Granular representation of the memory region containing a BitSlice.

BitDomainMut

Granular representation of the memory region containing a BitSlice.

Domain

Granular representation of the memory region containing a BitSlice.

DomainMut

Granular representation of the memory region containing a BitSlice.