Struct bitvec::index::BitMask

#[repr(transparent)]
pub struct BitMask<R> where
    R: BitRegister,  { /* fields omitted */ }

A multi-bit selection mask.

Unlike BitSel, which enforces a strict one-hot mask encoding, this mask type permits any number of bits to be set or unset. This is used to accumulate selections for a batch operation on a register.

Construction

It is only constructed by accumulating BitSel values. The chain of custody for safe construction in this module and in order ensures that all masks that are applied to register values can be trusted to not cause memory unsafety.

Implementations

impl<R> BitMask<R> where
R: BitRegister,

pub const ALL: Self

A full mask.

pub const ZERO: Self

An empty mask.

pub unsafe fn new(mask: R) -> Self

Wraps any R value as a bit-mask.

This constructor is provided to explicitly declare that an operation is discarding the numeric value of an integer and reading it only as a bit-mask.

Parameters

mask: Some integer value

Returns

mask wrapped as a bit-mask, with its numeric context discarded.

Safety

This function must only be called within a BitOrder::mask implementation which is verified to be correct.

Prefer accumulating BitSel values using the Sum implementation.

pub fn combine(self, sel: BitSel<R>) -> Self

Creates a new mask with a selector bit activated.

Parameters

self

sel: The selector bit to activate in the new mask.

Returns

A copy of self, with the selector at sel activated.

pub fn insert(&mut self, sel: BitSel<R>)

Inserts a selector into an existing mask.

Parameters

&mut self

sel: The selector bit to insert into the mask.

Effects

The selector’s bit in the self mask is activated.

pub fn test(self, sel: BitSel<R>) -> bool

Tests whether a mask contains a given selector bit.

Paramters

self

sel: The selector bit to test in the self mask.

Returns

Whether self has set the bit at sel.

pub fn value(self) -> R

Views the internal mask value.