Trait bitvec::mem::BitMemory

pub trait BitMemory: IsUnsigned + Sealed {
    const ONE: Self;
    const ALL: Self;
    const BITS: u8 = mem::size_of::() as u8 * 8;
    const INDX: u8 = Self::BITS.trailing_zeros() as u8;
    const MASK: u8 = Self::BITS - 1;
}

Description of an integer type.

This trait provides information used to describe integer-typed regions of memory and enables other parts of the crate to adequately describe the memory bus. This trait has no bearing on the processor instructions or registers used to interact with memory.

This trait cannot be implemented outside this crate.

Required Associated Constants

const ONE: Self

The value with only its least significant bit set to 1.

const ALL: Self

The value with all of its bits set to 1.

Provided Associated Constants

const BITS: u8 = mem::size_of::() as u8 * 8

The bit width of the integer.

mem::size_of returns the size in bytes, and bytes are always eight bits on architectures Rust targets.

Issue #76904 will place this constant on the fundamentals directly, as a u32.

const INDX: u8 = Self::BITS.trailing_zeros() as u8

The number of bits required to store an index in the range 0 .. BITS.

const MASK: u8 = Self::BITS - 1

A mask over all bits that can be used as an index within the element.