Trait bitvec::mem::BitMemory

pub trait BitMemory: IsUnsigned + Sealed {
    const INDX: u8 = Self::BITS.trailing_zeros() as u8;
    const MASK: u8 = Self::BITS as u8 - 1;
}

Description of an integer memory element.

This trait provides information used to describe integer-typed regions of memory and enables other parts of the project to adequately describe the memory bus. This trait has no bearing on the processor instructions or registers used to interact with memory. It solely describes integers that can exist on a system.

This trait cannot be implemented outside this crate.

Provided Associated Constants

const INDX: u8 = Self::BITS.trailing_zeros() as u8

The number of bits required to store an index in the range 0 .. BITS.

const MASK: u8 = Self::BITS as u8 - 1

A mask over all bits that can be used as an index within the element. This is the value with the least significant INDX-many bits set high.