Trait bitvec::mem::BitRegister

pub trait BitRegister: BitMemory + BitOps {
    const ONE: Self;
    const ALL: Self;
}

Description of a processor register.

This trait provides information used to describe processor registers. It only needs to contain constant values for 1 and !0; the rest of its information is contained in the presence or absence of its implementation on particular integers.

Required Associated Constants

const ONE: Self

The literal 1.

const ALL: Self

The literal !0.