Module bitvec::order

Ordering of bits within register elements.

bitvec data structures are parametric over any ordering of bits within a register type. The BitOrder trait translates a cursor position (indicated by the BitIdx type) to an electrical position (indicated by the BitPos type) within that register, thereby defining the order of traversal over a register.

Implementors of BitOrder are required to satisfy a set of requirements on their transform function, and must have identical behavior to the default-provided trait functions if they choose to override them for performance. These can all be proven by use of the verify or verify_for_type functions in the implementor’s test suite.

BitOrder is a stateless trait, and implementors should be zero-sized types.

!

Re-exports

pub use self::Lsb0 as LocalBits;

Structs

Lsb0

Traverses a register from the least significant bit to the most significant.

Msb0

Traverses a register from the most significant bit to the least significant.

Traits

BitOrder

An ordering over a register.

Functions

verify

Verifies a BitOrder implementation’s adherence to the stated rules.

verify_for_type

Verifies a BitOrder implementation’s adherence to the stated rules, for one register type.