Function bitvec::order::verify

pub fn verify<O>(verbose: bool) where
    O: BitOrder, 

Verifies a BitOrder implementation’s adherence to the stated rules.

This function checks some BitOrder implementation’s behavior on each of the BitRegister types it must handle, and reports any violation of the rules that it detects.

Type Parameters

O: The BitOrder implementation to test.

Parameters

verbose: Sets whether the test should print diagnostic information to stdout.

Panics

This panics if it detects any violation of the BitOrder implementation rules for O.