Module bitvec::prelude

Symbol Export

This module collects the general public API into a single place for bulk import, as use bitvec::prelude::*;, without polluting the root namespace of the crate.

This provides all the data structure types and macros, as well as the two traits needed to operate them as type parameters, by name. It also imports extension traits without naming them, so that their methods are available but their trait names are not.

Re-exports

pub use crate::array::BitArray;

pub use crate::field::BitField as _;

pub use crate::order::BitOrder;

pub use crate::order::LocalBits;

pub use crate::order::Lsb0;

pub use crate::order::Msb0;

pub use crate::slice::BitSlice;

pub use crate::store::BitStore;

pub use crate::view::AsBits;

pub use crate::view::AsMutBits;

pub use crate::view::BitView as _;

pub use crate::view::BitViewSized as _;

pub use crate::boxed::BitBox;

pub use crate::vec::BitVec;

Macros

BitArr

Bit-Array Type Definition

bitarr

Bit-Array Value Constructor

bitbox

Boxed Bit-Slice Constructor

bits

Bit-Slice Region Constructor

bitvec

Bit-Vector Constructor

Structs

BitPtr

Single-Bit Pointer

BitPtrRange

Bit-Pointer Range

BitRef

Proxy Bit-Reference