Module bitvec::slice

BitSlice Wide Reference

This module defines semantic operations on [u1], in contrast to the mechanical operations defined in BitPtr.

The &BitSlice handle has the same size and general layout as the standard Rust slice handle &[T]. Its binary layout is wholly incompatible with the layout of Rust slices, and must never be interchanged except through the provided APIs.

Structs

BitGuard
Write reference to a single bit.

BitSlice
A compact slice of bits, whose cursor and storage types can be customized.

Chunks
State keeper for chunked iteration over a BitSlice.

ChunksExact
State keeper for exact chunked iteration over a BitSlice.

ChunksExactMut
State keeper for mutable exact chunked iteration over a BitSlice.

ChunksMut
State keeper for mutable chunked iteration over a BitSlice.

Iter
State keeper for iteration over a BitSlice.

RChunks
State keeper for reverse chunked iteration over a BitSlice.

RChunksExact
State keeper for reverse exact iteration over a BitSlice.

RChunksExactMut
State keeper for mutable reverse exact chunked iteration over a BitSlice.

RChunksMut
State keeper for mutable reverse chunked iteration over a BitSlice.

Windows
State keeper for sliding-window iteration over a BitSlice.