Function bitvec::slice::from_raw_parts_mut

pub unsafe fn from_raw_parts_mut<'a, O, T>(
    data: *mut T,
    len: usize
) -> &'a mut BitSlice<O, T>where
    O: BitOrder,
    T: 'a + BitStore,

Performs the same functionality as from_raw_parts, except that a mutable slice is returned.

This function is unsafe for the same reason as from_raw_parts, as well as not being able to provide a non-aliasing guarantee of the returned mutable slice. data must be non-null and aligned even for zero-length slices as with from_raw_parts. The total size of the slice must be no larger than BitPtr::<T>::MAX_ELTS elements in memory.

See the documentation of from_raw_parts for more details.

Safety

See from_raw_parts.

Original

core::slice::from_raw_parts_mut