Module bitvec::vec

A dynamically-allocated buffer containing a BitSlice<O, T> region.

You can read the standard library’s alloc::vec module documentation here.

This module defines the BitVec buffer, and all of its associated support code.

BitVec is equivalent to Vec<bool>, in its operation and in its relationship to the BitSlice type. Most of the interesting work to be done on a bit-sequence is implemented in BitSlice, to which BitVec dereferences, and the vector container itself only exists to maintain ownership, implement dynamic resizing, and provide some specializations that cannot safely be done on BitSlice alone.

!

Structs

BitVec

A vector of individual bits, allocated on the heap.

Drain

A draining iterator for BitVec<O, T>.

IntoIter

An iterator that moves out of a vector.

Splice

A splicing iterator for BitVec.