Struct bitvec::vec::BitVec

#[repr(C)]
pub struct BitVec<O = Lsb0, T = usize> where
    O: BitOrder,
    T: BitStore,  { /* private fields */ }

A vector of individual bits, allocated on the heap.

This is a managed, heap-allocated, buffer that contains a BitSlice region. It is analagous to Vec<bool>, and is written to be as close as possible to drop-in replacabale for it. This type contains little interesting behavior in its own right, dereferencing instead to BitSlice for manipulation of the buffer contents, and serves primarily as an interface to the allocator. If you require statically-allocated, fixed-size, owned buffers, you should use the BitArray type.

Because BitVec directly owns its memory, and can guarantee that no other object in a program has access to its buffers, BitVec is able to override some behavior from BitSlice in more efficient manners.

Documentation

All APIs that mirror something in the standard library will have an Original section linking to the corresponding item. All APIs that have a different signature or behavior than the original will have an API Differences section explaining what has changed, and how to adapt your existing code to the change.

These sections look like this:

Original

Vec<T>

API Differences

The buffer type Vec<bool> has no type parameters. BitVec<O, T> has the same two type parameters as BitSlice<O, T>. Otherwise, BitVec is able to implement the full API surface of Vec<bool>.

Behavior

Because BitVec is a fully-owned buffer, it is able to operate on its memory without concern for any other views that may alias. This enables it to specialize some BitSlice behavior to be faster or more efficient.

Type Parameters

This takes the same two type parameters, O: BitOrder and T: BitStore, as BitSlice.

Safety

Like BitSlice, BitVec is exactly equal in size to Vec, and is also absolutely representation-incompatible with it. You must never attempt to type-cast between Vec<T> and BitVec in any way, nor attempt to modify the memory value of a BitVec handle. Doing so will cause allocator and memory errors in your program, likely inducing a panic.

Everything in the BitVec public API, even the unsafe parts, are guaranteed to have no more unsafety than their equivalent items in the standard library. All unsafe APIs will have documentation explicitly detailing what the API requires you to uphold in order for it to function safely and correctly. All safe APIs will do so themselves.

Performance

The choice of T: BitStore type parameter can impact your vector’s performance, as the allocator operates in units of T rather than in bits. This means that larger register types will increase the amount of memory reserved in each call to the allocator, meaning fewer calls to .push() will actually cause a reällocation. In addition, iteration over the vector is governed by the BitSlice characteristics on the type parameter. You are generally better off using larger types when your vector is a data collection rather than a specific I/O protocol buffer.

Macro Construction

Heap allocation can only occur at runtime, but the bitvec! macro will construct an appropriate BitSlice buffer at compile-time, and at run-time, only copy the buffer into a heap allocation.

Implementations

impl<O, T> BitVec<O, T> where
O: BitOrder,
T: BitStore,

pub fn new() -> Self

Constructs a new, empty BitVec<O, T>.

The vector will not allocate until bits are pushed into it.

Original

Vec::new

Examples

use bitvec::prelude::*;

let mut bv = BitVec::<LocalBits, usize>::new();

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty BitVec<O, T> with the specified capacity.

The vector will be able to hold at least capacity bits without reällocating. If capacity is 0, the vector will not allocate.

It is important to note that although the returned vector has the capacity specified, the vector will have a zero length. For an explanation of the difference between length and capacity, see Capacity and reällocation.

Original

Vec::with_capacity

Panics

Panics if the requested capacity exceeds the vector’s limits.

Examples

use bitvec::prelude::*;

let mut bv = BitVec::<LocalBits, usize>::with_capacity(10);

// The vector contains no items, even though it has capacity for more
assert_eq!(bv.len(), 0);

// These are all done without reallocating...
for i in 0..10 {
  bv.push(true);
}

// ...but this may make the vector reallocate
bv.push(false);

pub unsafe fn from_raw_parts(
pointer: *mut BitSlice<O, T>,
capacity: usize
) -> Self

Creates a BitVec<O, T> directly from the raw components of another bit-vector.

Original

Vec::from_raw_parts

API Differences

Ordinary vectors decompose into their buffer pointer and element length separately; bit vectors must keep these two components bundled into the *BitSlice region pointer. As such, this only accepts two components; the slice pointer and the buffer capacity.

Vec could define its raw parts as *[T] and usize also, but Rust does not make working with raw slice pointers easy.

Panics

This function panics if pointer is the null pointer.

Safety

This is highly unsafe, due to the number of invariants that aren’t checked:

pointer needs to have been previously allocated via BitVec<O, T> (at least, it’s highly likely to be incorrect if it wasn’t).

T needs to have the same size and alignment as what pointer was allocated with. (T having a less strict alignment is not sufficient; the alignment really needs to be equal to satisfy the [dealloc] requirement that memory must be allocated and deällocated with the same layout.)

capacity needs to be the capacity that the pointer was allocated with.

In addition to the invariants inherited from Vec::from_raw_parts, the fact that this function takes a bit-slice pointer adds another one:

pointer MUST NOT have had its value modified in any way in the time when it was outside of a bitvec container type.

Violating these will cause problems like corrupting the allocator’s internal data structures. For example it is not safe to build a BitVec<_, u8> from a pointer to a C char array with length size_t. It’s also not safe to build one from a BitVec<_, u16> and its length, becauset the allocator cares about the alignment, and these two types have different alignments. The buffer was allocated with alignment 2 (for u16), but after turning it into a BitVec<_, u8>, it’ll be deällocated with alignment 1.

The ownership of pointer is effectively transferred to the BitVec<O, T> which may then deällocate, reällocate, or change the contents of memory pointed to by the pointer at will. Ensure that nothing else uses the pointer after calling this function.

Examples

use bitvec::prelude::*;
use bitvec as bv;
use core::mem;

let bv = bitvec![0, 1, 0, 1];

// Prevent running `bv`’s destructor so we are in complete control
// of the allocation.
let mut bv = mem::ManuallyDrop::new(bv);

// Pull out the various important pieces of information about `bv`
let p = bv.as_mut_ptr();
let e = bv.elements();
let cap = bv.capacity();

unsafe {
  let bits = bv::slice::from_raw_parts_mut::<LocalBits, _>(p, e);
  let len = bits.len();

  // Overwrite memory with a new pattern
  bits.iter_mut().for_each(|mut b| *b = true);

  // Put everything back together into a BitVec
  let rebuilt = BitVec::from_raw_parts(bits as *mut _, cap);
  assert_eq!(rebuilt.len(), len);
}

pub fn capacity(&self) -> usize

Returns the number of bits the vector can hold without reällocating.

Original

Vec::capacity

Examples

use bitvec::prelude::*;

let bv: BitVec<LocalBits, usize> = BitVec::with_capacity(100);
assert!(bv.capacity() >= 100);

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more bits to be inserted in the given BitVec<O, T>. The collection may reserve more space to avoid frequent reällocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

Original

Vec::reserve

Panics

Panics if the new capacity exceeds the vector’s limits.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1];
bv.reserve(100);
assert!(bv.capacity() >= 101);

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more bits to be inserted in the given BitVec<O, T>. After calling reserve_exact, capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore, capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Original

Vec::reserve_exact

Panics

Panics if the new capacity exceeds the vector’s limits.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1];
bv.reserve_exact(100);
assert!(bv.capacity() >= 101);

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the vector as much as possible.

It will drop down as close as possible to the length but the allocator may still inform the vector that there is space for a few more bits.

Original

Vec::shrink_to_fit

Examples

use bitvec::prelude::*;

let mut bv = BitVec::<LocalBits, usize>::with_capacity(100);
bv.extend([false, true, false].iter().copied());
assert!(bv.capacity() >= 100);
bv.shrink_to_fit();
assert!(bv.capacity() >= 3);

pub fn into_boxed_slice(self) -> Box<[T]>

Converts the vector into [Box<[T]>].

Note that this will drop any excess capacity.

Original

Vec::into_boxed_slice

Analogue

See [into_boxed_bitslice] for a BitVec -> BitBox transform.

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0];

let slice = bv.into_boxed_slice();
assert_eq!(slice.len(), 1);

Any excess capacity is removed:

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(100);
bv.extend([false, true, false].iter().copied());

assert!(bv.capacity() >= 100);
let slice = bv.into_boxed_slice();
assert_eq!(slice.into_vec().capacity(), 1);

[Box<[T]>]: https://doc.rust-lang.org/alloc/boxed/struct.Box.html [into_boxed_bitslice]: #method.into_boxed_bitslice

pub fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len bits and dropping the rest.

If len is greater than the vector’s current length, this has no effect.

The drain method can emulate truncate, but causes the excess bits to be returned instead of dropped.

Note that this method has no effect on the allocated capacity of the vector, nor does it erase truncated memory. Bits in the allocated memory that are outside of the .as_bitslice() view always have unspecified values, and cannot be relied upon to be zero.

Original

Vec::truncate

Examples

Truncating a five bit vector to two bits:

use bitvec::prelude::*;

let mut bv = bitvec![1; 5];
bv.truncate(2);
assert_eq!(bv.len(), 2);
assert!(bv.as_slice()[0].count_ones() >= 5);

No truncation occurs when len is greater than the vector’s current length:

use bitvec::prelude::*;

let mut bv = bitvec![1; 3];
bv.truncate(8);
assert_eq!(bv.len(), 3);

Truncating when len == 0 is equivalent to calling the clear method.

use bitvec::prelude::*;

let mut bv = bitvec![0; 3];
bv.truncate(0);
assert!(bv.is_empty());

pub fn as_slice(&self) -> &[T]

Extracts an element slice containing the entire vector.

Original

Vec::as_slice

Analogue

See as_bitslice for a &BitVec -> &BitSlice transform.

Examples

use bitvec::prelude::*;
use std::io::{self, Write};
let buffer = bitvec![Msb0, u8; 0, 1, 0, 1, 1, 0, 0, 0];
io::sink().write(buffer.as_slice()).unwrap();

pub fn as_mut_slice(&mut self) -> &mut [T]

Extracts a mutable slice of the entire vector.

Original

Vec::as_mut_slice

Analogue

See as_mut_bitslice for a &mut BitVec -> &mut BitSlice transform.

Examples

use bitvec::prelude::*;
use std::io::{self, Read};
let mut buffer = bitvec![Msb0, u8; 0; 24];
io::repeat(0b101).read_exact(buffer.as_mut_slice()).unwrap();

pub fn as_ptr(&self) -> *const T

Returns a raw pointer to the vector’s buffer.

The caller must ensure that the vector outlives the pointer this function returns, or else it will end up pointing to garbage. Modifying the vector may cause its buffer to be reällocated, which would also make any pointers to it invalid.

The caller must also ensure that the memory the pointer (non-transitively) points to is never written to (except inside an UnsafeCell) using this pointer or any pointer derived from it. If you need to mutate the contents of the slice, use [as_mut_ptr].

Original

Vec::as_ptr

Analogue

See [as_bitptr] for a &BitVec -> *const BitSlice transform.

Examples

use bitvec::prelude::*;

let bv = bitvec![Lsb0; 0, 1, 0, 1];
let bv_ptr = bv.as_ptr();

unsafe {
  assert_eq!(*bv_ptr, 0b1010);
}

pub fn as_mut_ptr(&mut self) -> *mut T

Returns an unsafe mutable pointer to the vector’s buffer.

The caller must ensure that the vector outlives the pointer this function returns, or else it will end up pointing to garbage. Modifying the vector may cause its buffer to be reällocated, which would also make any pointers to it invalid.

Original

Vec::as_mut_ptr

Analogue

See as_mut_bitptr for a &mut BitVec -> *mut BitSlice transform.

Eaxmples

use bitvec::prelude::*;

let size = 4;
let mut bv: BitVec<Msb0, usize> = BitVec::with_capacity(size);
let bv_ptr = bv.as_mut_ptr();

unsafe {
  *bv_ptr = !0;
  bv.set_len(size);
}
assert_eq!(bv.len(), 4);
assert!(bv.all());

pub unsafe fn set_len(&mut self, new_len: usize)

Forces the length of the vector to new_len.

This is a low-level operation that maintains none of the normal invariants of the type. Normally changing the length of a vector is done using one of the safe operations instead, such as truncate, resize, extend, or clear.

Original

Vec::set_len

Safety

new_len must be less than or equal to capacity().

Examples

This method can be useful for situations in which the vector is serving as a buffer for other code, particularly over FFI:

use bitvec::prelude::*;

// `bitvec` could pair with `rustler` for a better bitstream
type ErlBitstring = BitVec<Msb0, u8>;
let mut bits_read = 0;
// An imaginary Erlang function wants a large bit buffer.
let mut buf = ErlBitstring::with_capacity(32_768);
// SAFETY: When `erl_read_bits` returns `ERL_OK`, it holds that:
// 1. `bits_read` bits were initialized.
// 2. `bits_read` <= the capacity (32_768)
// which makes `set_len` safe to call.
unsafe {
  // Make the FFI call...
  let status = erl_read_bits(&mut buf, 10, &mut bits_read);
  if status == ERL_OK {
    // ...and update the length to what was read in.
    buf.set_len(bits_read);
  }
}

pub fn swap_remove(&mut self, index: usize) -> bool

Removes a bit from the vector and returns it.

The removed bit is replaced by the last bit of the vector.

This does not preserve ordering, but is O(1).

Original

Vec::swap_remove

Panics

Panics if index is out of bounds.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1, 0, 1];
assert!(!bv.swap_remove(1));
assert_eq!(bv, bits![0, 1, 1, 0]);

assert!(!bv.swap_remove(0));
assert_eq!(bv, bits![0, 1, 1]);

pub fn insert(&mut self, index: usize, value: bool)

Inserts a bit at position index within the vector, shifting all bits after it to the right.

Original

Vec::insert

Panics

Panics if index > len.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0; 5];
bv.insert(4, true);
assert_eq!(bv, bits![0, 0, 0, 0, 1, 0]);
bv.insert(2, true);
assert_eq!(bv, bits![0, 0, 1, 0, 0, 1, 0]);

pub fn remove(&mut self, index: usize) -> bool

Removes and returns the bit at position index within the vector, shifting all bits after it to the left.

Original

Vec::remove

Panics

Panics if index is out of bounds.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0];
assert!(bv.remove(1));
assert_eq!(bv, bits![0, 0]);

pub fn retain<F>(&mut self, func: F) where
F: FnMut(usize, &bool) -> bool,

Retains only the bits specified by the predicate.

In other words, remove all bits b such that func(idx(b), &b) returns false. This method operates in place, visiting each bit exactly once in the original order, and preserves the order of the retained bits.

Original

Vec::retain

API Differences

In order to allow more than one bit of information for the split decision, the predicate receives the index of each bit, as well as its value.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 1, 0, 0, 1];
bv.retain(|i, b| (i % 2 == 0) ^ b);
assert_eq!(bv, bits![0, 1, 0, 1]);

pub fn push(&mut self, value: bool)

Appends a bit to the back of a collection.

Original

Vec::push

Panics

Panics if the number of bits in the vector exceeds the maximum vector capacity.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0];
bv.push(true);
assert_eq!(bv.count_ones(), 1);

pub fn pop(&mut self) -> Option<bool>

Removes the last bit from a vector and returns it, or None if it is empty.

Original

Vec::pop

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1];
assert_eq!(bv.pop(), Some(true));
assert!(bv.not_any());

pub fn append<O2, T2>(&mut self, other: &mut BitVec<O2, T2>) where
O2: BitOrder,
T2: BitStore,

Moves all the bits of other into self, leaving other empty.

Original

Vec::append

Panics

Panics if the number of bits overflows the maximum vector capacity.

Examples

use bitvec::prelude::*;

let mut bv1 = bitvec![0; 10];
let mut bv2 = bitvec![1; 10];

bv1.append(&mut bv2);

assert_eq!(bv1.count_ones(), 10);
assert!(bv2.is_empty());

pub fn drain<R>(&mut self, range: R) -> Drain<'_, O, T> where
R: RangeBounds<usize>,

Creates a draining iterator that removes the specified range in the vector and yields the removed items.

Note 1: The bit range is removed even if the iterator is only partially consumed or not consumed at all.

Note 2: It is unspecified how many bits are removed from the vector if the Drain value is leaked.

Original

Vec::drain

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the vector.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 1];
let bv2: BitVec = bv.drain(1 ..).collect();
assert_eq!(bv, bits![0]);
assert_eq!(bv2, bits![1, 1]);

// A full range clears the vector
bv.drain(..);
assert_eq!(bv, bits![]);

pub fn clear(&mut self)

Clears the vector, removing all values.

Note that this method has no effect on the allocated capacity of the vector.

Original

Vec::clear

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 1];

bv.clear();

assert!(bv.is_empty());

pub fn split_off(&mut self, at: usize) -> Self

Splits the collection into two at the given index.

Returns a newly allocated vector containing the elements in range [at, len). After the call, the original vector will be left containing the bits [0, at) with its previous capacity unchanged.

Original

Vec::split_off

Panics

Panics if at > len.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1];
let bv2 = bv.split_off(1);
assert_eq!(bv, bits![0]);
assert_eq!(bv2, bits![0, 1]);

pub fn resize_with<F>(&mut self, new_len: usize, func: F) where
F: FnMut() -> bool,

Resizes the BitVec in-place so that len is equal to new_len.

If new_len is greater than len, the BitVec is extended by the difference, with each additional slot filled with the result of calling the closure func. The return values from func will end up in the BitVec in the order they have been generated.

If new_len is less than len, the Vec is simply truncated.

This method uses a closure to create new values on every push. If you’d rather Clone a given bit, use resize. If you want to use the Default trait to generate values, you can pass Default::default as the second argument.

Original

Vec::resize_with

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 3];
bv.resize_with(5, Default::default);
assert_eq!(bv, bits![1, 1, 1, 0, 0]);

let mut bv = bitvec![];
let mut p = 0;
bv.resize_with(4, || { p += 1; p % 2 == 0 });
assert_eq!(bv, bits![0, 1, 0, 1]);

pub fn resize(&mut self, new_len: usize, value: bool)

Resizes the BitVec in-place so that len is equal to new_len.

If new_len is greater than len, the BitVec is extended by the difference, with each additional slot filled with value. If new_len is less than len, the BitVec is simply truncated.

This method requires a single bool value. If you need more flexibility, use resize_with.

Original

Vec::resize

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1];
bv.resize(3, false);
assert_eq!(bv, bits![1, 0, 0]);

let mut bv = bitvec![1; 4];
bv.resize(2, false);
assert_eq!(bv, bits![1; 2]);

pub fn extend_from_slice(&mut self, other: &[bool])

Clones and appends all bools in a slice to the BitVec.

Iterates over the slice other, clones each bool, and then appends it to the BitVec. The other slice is traversed in-order.

Prefer the Extend implementation; this method is retained only for API compatibility, and offers no performance benefit.

Original

Vec::extend_from_slice

Analogue

See [extend_from_bitslice] for the method to append a bit-slice of the same type parameters to a bit-vector.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0];
bv.extend_from_slice(&[true]);
assert_eq!(bv, bits![0, 1]);

[extend]: #impl-Extend<%26’a bool> [extend_from_bitslice]: #method.extend_from_bitslice

pub fn splice<R, I>(
&mut self,
range: R,
replace_with: I
) -> Splice<'_, O, T, I::IntoIter> where
R: RangeBounds<usize>,
I: IntoIterator<Item = bool>,

Creates a splicing iterator that replaces the specified range in the vector with the given replace_with iterator and yields the removed items. replace_with does not need to be the same length as range.

The element range is removed even if the iterator is not consumed until the end.

It is unspecified how many bits are removed from the vector if the Splice value is leaked.

The input iterator replace_with is only consumed when the Splice value is dropped.

This is optimal if:

the tail (bits in the vector after range) is empty

or replace_with yields fewer bits than range’s length

or the lower bound of its size_hint() is exact

Otherwise, a temporary vector is allocated and the tail is moved twice.

Original

Vec::splice

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0];
let new = bits![1, 0];
let old: BitVec = bv.splice(.. 2, new.iter().copied()).collect();
assert_eq!(bv, bits![1, 0, 0]);
assert_eq!(old, bits![0, 1]);

impl<O, T> BitVec<O, T> where
O: BitOrder,
T: BitStore,

Methods specific to BitVec<_, T>, and not present on Vec<T>.

pub fn repeat(bit: bool, len: usize) -> Self

Constructs a BitVec from a value repeated many times.

This function is equivalent to the bitvec![O, T; bit; len] macro call, and is in fact the implementation of that macro syntax.

Parameters

bit: The bit value to which all len allocated bits will be set.

len: The number of live bits in the constructed BitVec.

Returns

A BitVec with len live bits, all set to bit.

Examples

use bitvec::prelude::*;

let bv = BitVec::<Msb0, u8>::repeat(true, 20);
assert_eq!(bv, bits![1; 20]);

pub fn from_bitslice(slice: &BitSlice<O, T>) -> Self

Clones a &BitSlice into a BitVec.

Original

<Vec<T: Clone> as Clone>::clone

Effects

This performs a direct element-wise copy from the source slice to the newly-allocated buffer, then sets the vector to have the same starting bit as the slice did. This allows for faster behavior. If you require that the vector start at the leading edge of the first element, use force_align to guarantee this.

Examples

use bitvec::prelude::*;

let bits = bits![0, 1, 0, 1, 1, 0, 1, 1];
let bv = BitVec::from_bitslice(&bits[2 ..]);
assert_eq!(bv, bits[2 ..]);

pub fn from_vec(vec: Vec<T>) -> Self

Converts a Vec<T> into a BitVec<O, T> without copying its buffer.

Parameters

vec: A vector to view as bits.

Returns

A BitVec over the vec buffer.

Panics

This panics if vec is too long to convert into a BitVec. See BitSlice::MAX_ELTS.

Examples

use bitvec::prelude::*;

let vec = vec![0u8; 4];
let bv = BitVec::<LocalBits, _>::from_vec(vec);
assert_eq!(bv, bits![0; 32]);

pub fn try_from_vec(vec: Vec<T>) -> Result<Self, Vec<T>>

Converts a Vec<T> into a BitVec<O, T> without copying its buffer.

This method takes ownership of a memory buffer and enables it to be used as a bit-vector. Because Vec can be longer than BitVecs, this is a fallible method, and the original vector will be returned if it cannot be converted.

Parameters

vec: Some vector of memory, to be viewed as bits.

Returns

If vec is short enough to be viewed as a BitVec, then this returns a BitVec over the vec buffer. If vec is too long, then this returns vec unmodified.

Examples

use bitvec::prelude::*;

let vec = vec![0u8; 4];
let bv = BitVec::<LocalBits, _>::try_from_vec(vec).unwrap();
assert_eq!(bv, bits![0; 32]);

An example showing this function failing would require an allocation exceeding !0usize >> 3 bytes in size, which is infeasible to produce.

pub fn extend_from_bitslice(&mut self, other: &BitSlice<O, T>)

Copies all bits in a BitSlice into the BitVec.

This is provided for API completeness; it has no performance benefits compared to use of the Extend implementation.

Parameters

&mut self

other: A BitSlice reference of the same type parameters as self.

Behavior

self is extended by the length of other, and then the contents of other are copied into the newly-allocated end of self.

use bitvec::prelude::*;

let mut bv = bitvec![0, 1];
bv.extend_from_bitslice(bits![1, 1, 0, 1]);

assert_eq!(bv, bits![0, 1, 1, 1, 0, 1]);

Extend: #impl-Extend<%26’a bool> [.as_bitslice()]: #method.as_bitslice()

pub fn into_boxed_bitslice(self) -> BitBox<O, T>

Converts the vector into BitBox<O, T>.

Note that this will drop any excess capacity.

Original

Vec::into_boxed_slice

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 50];
let bb: BitBox = bv.into_boxed_bitslice();
assert_eq!(bb, bits![1; 50]);

pub fn into_vec(self) -> Vec<T>

Converts the vector back into an ordinary vector of memory elements.

This does not affect the vector’s buffer, only the handle used to control it.

Parameters

self

Returns

An ordinary vector containing all of the bit-vector’s memory buffer.

Examples

use bitvec::prelude::*;

let bv = bitvec![0; 5];
let vec = bv.into_vec();
assert_eq!(vec, [0]);

pub fn elements(&self) -> usize

Gets the number of elements T that contain live bits of the vector.

Examples

use bitvec::prelude::*;

let bv = bitvec![LocalBits, u16; 1; 50];
assert_eq!(bv.elements(), 4);

pub fn set_uninitialized(&mut self, value: bool)

Sets the uninitialized bits of the vector to a fixed value.

This method modifies all bits in the allocated buffer that are outside the self.as_bitslice() view so that they have a consistent value. This can be used to zero the uninitialized memory so that when viewed as a raw memory slice, bits outside the live region have a predictable value.

Examples

use bitvec::prelude::*;

let mut bv = 220u8.view_bits::<Lsb0>().to_bitvec();
assert_eq!(bv.as_slice(), &[220u8]);
bv.truncate(4);
assert_eq!(bv.count_ones(), 2);
assert_eq!(bv.as_slice(), &[220u8]);

bv.set_uninitialized(false);
assert_eq!(bv.as_slice(), &[12u8]);

bv.set_uninitialized(true);
assert_eq!(bv.as_slice(), &[!3u8]);

pub fn force_align(&mut self)

Ensures that the live region of the vector’s contents begins at the leading edge of the buffer.

Examples

use bitvec::prelude::*;

let data = 0x3Cu8;
let bits = data.view_bits::<Msb0>();

let mut bv = bits[2 .. 6].to_bitvec();
assert_eq!(bv, bits[2 .. 6]);
assert_eq!(bv.as_slice()[0], data);

bv.force_align();
assert_eq!(bv, bits[2 .. 6]);
//  It is not specified what happens to bits that are no longer used.
assert_eq!(bv.as_slice()[0] & 0xF0, 0xF0);

pub fn set_elements(&mut self, element: T::Mem)

Writes a value into every element that the vector considers live.

This unconditionally writes element into each live location in the backing buffer, without altering the BitVec’s length or capacity.

It is unspecified what effects this has on the allocated but dead elements in the buffer.

Parameters

&mut self

element: The value which will be written to each live location in the vector’s buffer.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![LocalBits, u8; 0; 10];
assert_eq!(bv.as_slice(), [0, 0]);
bv.set_elements(0xA5);
assert_eq!(bv.as_slice(), [0xA5, 0xA5]);

pub fn as_bitslice(&self) -> &BitSlice<O, T>

Views the buffer’s contents as a BitSlice.

This is equivalent to &bv[..].

Original

Vec::as_slice

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 1, 0];
let bits = bv.as_bitslice();

pub fn as_mut_bitslice(&mut self) -> &mut BitSlice<O, T>

Extracts a mutable bit-slice of the entire vector.

Equivalent to &mut bv[..].

Original

Vec::as_mut_slice

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 1];
let bits = bv.as_mut_bitslice();
bits.set(0, true);

pub fn as_bitptr(&self) -> *const BitSlice<O, T>

Returns a raw pointer to the vector’s region.

The caller must ensure that the vector outlives the pointer this function returns, or else it will end up pointing to garbage. Modifying the vector may cause its buffer to be reallocated, which would also make any pointers to it invalid.

The caller must also ensure that the memory the pointer (non-transitively) points to is never written to (except inside an UnsafeCell) using this pointer or any pointer derived from it. If you need to mutate the contents of the region, use as_mut_bitptr.

This pointer is an opaque crate-internal type. Its in-memory representation is unsafe to modify in any way. The only safe action to take with this pointer is to pass it, unchanged, back into a bitvec API.

Examples

use bitvec::prelude::*;

let bv = bitvec![0; 20];
let ptr = bv.as_bitptr();

let bits = unsafe { &*ptr };
assert_eq!(bv, bits);

pub fn as_mut_bitptr(&mut self) -> *mut BitSlice<O, T>

Returns an unsafe mutable pointer to the vector’s region.

The caller must ensure that the vector outlives the pointer this function returns, or else it will end up pointing to garbage. Modifying the vector may cause its buffer to be reallocated, which would also make any pointers to it invalid.

This pointer is an opaque crate-internal type. Its in-memory representation is unsafe to modify in any way. The only safe action to take with this pointer is to pass it, unchanged, back into a bitvec API.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0; 20];
let ptr = bv.as_mut_bitptr();

let bits = unsafe { &mut *ptr };
assert_eq!(bv, bits);