Struct bitvec::vec::BitVec

#[repr(C)]pub struct BitVec<O = Local, T = usize> where
    O: BitOrder,
    T: BitStore,  { /* fields omitted */ }

A compact Vec of bits, whose order and storage type can be customized.

BitVec is a newtype wrapper over Vec, and as such is exactly three words in size on the stack.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();
bv.push(false);
bv.push(true);

assert_eq!(bv.len(), 2);
assert_eq!(bv[0], false);

assert_eq!(bv.pop(), Some(true));
assert_eq!(bv.len(), 1);

bv.set(0, true);
assert_eq!(bv[0], true);

bv.extend([0u8, 1, 0].iter().map(|n| *n != 0u8));
for bit in &*bv {
  println!("{}", bit);
}
assert_eq!(bv, bitvec![1, 0, 1, 0]);

The bitvec! macro is provided to make initialization more convenient.

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 2, 3];
bv.push(false);
assert_eq!(bv, bitvec![0, 1, 1, 1, 0]);

It can also initialize each element of a BitVec<_, T> with a given value. This may be more efficient than performing allocation and initialization in separate steps, especially when initializing a vector of zeros:

use bitvec::prelude::*;

let bv = bitvec![0; 15];
assert_eq!(bv, bitvec![0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]);

// The following is equivalent, but potentially slower:
let mut bv1: BitVec = BitVec::with_capacity(15);
bv1.resize(15, false);

Use a BitVec<T> as an efficient stack:

use bitvec::prelude::*;
let mut stack: BitVec = BitVec::new();

stack.push(false);
stack.push(true);
stack.push(true);

while let Some(top) = stack.pop() {
  //  Prints true, true, false
  println!("{}", top);
}

Indexing

The BitVec type allows you to access values by index, because it implements the Index trait. An example will be more explicit:

use bitvec::prelude::*;

let bv = bitvec![0, 0, 1, 1];
println!("{}", bv[1]); // it will display 'false'

However, be careful: if you try to access an index which isn’t in the BitVec, your software will panic! You cannot do this:

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0, 1];
println!("{}", bv[6]); // it will panic!

In conclusion: always check if the index you want to get really exists before doing it.

Slicing

A BitVec is growable. A BitSlice, on the other hand, is fixed size. To get a bit slice, use &. Example:

use bitvec::prelude::*;
fn read_bitslice(slice: &BitSlice) {
    // use slice
}

let bv = bitvec![0, 1];
read_bitslice(&bv);

// … and that’s all!
// you can also do it like this:
let bs: &BitSlice = &bv;

In Rust, it’s more common to pass slices as arguments rather than vectors when you do not want to grow or shrink it. The same goes for Vec and [&[]], and String and &str.

Capacity and reallocation

The capacity of a bit vector is the amount of space allocated for any future bits that will be added onto the vector. This is not to be confused with the length of a vector, which specifies the number of live, useful bits within the vector. If a vector’s length exceeds its capacity, its capacity will automatically be increased, but its storage elements will have to be reallocated.

For example, a bit vector with capacity 10 and length 0 would be an allocated, but uninhabited, vector, with space for ten more bits. Pushing ten or fewer bits onto the vector will not change its capacity or cause reallocation to occur. However, if the vector’s length is increased to eleven, it will have to reallocate, which can be slow. For this reason, it is recommended to use BitVec::with_capacity whenever possible to specify how big the bit vector is expected to get.

Guarantees

Due to its incredibly fundamental nature, BitVec makes a lot of guarantees about its design. This ensures that it is as low-overhead as possible in the general case, and can be correctly manipulated in fundamental ways by unsafe code.

Most fundamentally, BitVec is and always will be a ([BitPtr], capacity) doublet. No more, no less. The order of these fields is unspecified, and you should only interact with the members through the provided APIs. Note that BitPtr is not directly manipulable, and must never be written or interpreted as anything but opaque binary data by user code.

When a BitVec has allocated memory, then the memory to which it points is on the heap (as defined by the allocator Rust is configured to use by default), and its pointer points to len initialized bits in order of the BitOrder type parameter, followed by capacity - len logically uninitialized bits.

BitVec will never perform a “small optimization” where elements are stored in its handle representation, for two reasons:

It would make it more difficult for user code to correctly manipulate a BitVec. The contents of the BitVec would not have a stable address if the handle were moved, and it would be more difficult to determine if a BitVec had allocated memory.

It would penalize the general, heap-allocated, case by incurring a branch on every access.

BitVec will never automatically shrink itself, even if it is emptied. This ensures that no unnecessary allocations or deallocations occur. Emptying a BitVec and then refilling it to the same length will incur no calls to the allocator. If you wish to free up unused memory, use shrink_to_fit.

Erasure

BitVec will not specifically overwrite any data that is removed from it, nor will it specifically preserve it. Its uninitialized memory is scratch space that may be used however the implementation desires, and must not be relied upon as stable. Do not rely on removed data to be erased for security purposes. Even if you drop a BitVec, its buffer may simply be reused for other data structures in your program. Even if you zero a BitVec’s memory first, that may not actually occur if the optimizer does not consider this an observable side effect. There is one case that will never break, however: using unsafe to construct a [T] slice over the BitVec’s capacity, and writing to the excess space, then increasing the length to match, is always valid.

Type Parameters

O: BitOrder: An implementor of the BitOrder trait. This type is used to convert semantic indices into concrete bit positions in elements, and store or retrieve bit values from the storage type.

T: BitStore: An implementor of the BitStore trait: u8, u16, u32, or u64 (64-bit systems only). This is the actual type in memory that the vector will use to store data.

Safety

The BitVec handle has the same size as standard Rust Vec handles, but it is extremely binary incompatible with them. Attempting to treat BitVec<_, T> as Vec<T> in any manner except through the provided APIs is catastrophically unsafe and unsound.

[&[]]: https://doc.rust-lang.org/stable/std/primitive.slice.html

Implementations

impl<O, T> BitVec<O, T> where
O: BitOrder,
T: BitStore,

pub fn new() -> Self

Constructs a new, empty BitVec<C, T>.

The vector will not allocate until elements are pushed onto it.

let mut bv: BitVec<Local, usize> = BitVec::new();

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty BitVec<C, T> with the specified capacity.

The vector will be able to hold at least capacity bits without reallocating. If capacity is 0, the vector will not allocate.

It is important to note that although the returned vector has the capacity specified, the vector will have a zero length. For an explanation of the difference between length and capacity, see Capacity and reallocation.

pub fn capacity(&self) -> usize

Returns the number of bits the vector can hold without reallocating.

Examples

let bv: BitVec<Local, usize> = BitVec::with_capacity(100);
assert!(bv.capacity() >= 100);

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more bits to be inserted in the given BitVec<C, T>. The collection may reserve more space to avoid frequent reallocations. After calling reserve, the capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Panics

Panics if the new capacity overflows BitPtr::<T>::MAX_BITS.

Examples

let mut bv = bitvec![1];
bv.reserve(10);
assert!(bv.capacity() >= 11);

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more bits to be inserted in the given BitVec<C, T>. After calling reserve_exact, capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore, capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Panics

Panics if the new capacity overflows BitPtr::<T>::MAX_BITS.

Examples

let mut bv = bitvec![1];
bv.reserve_exact(10);
assert!(bv.capacity() >= 11);

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the vector as much as possible.

It will drop down as close as possible to the length but the allocator may still inform the vector that there is space for a few more elements.

Examples

let mut bv: BitVec<Local, usize> = BitVec::with_capacity(10);
bv.extend([true, false, true].iter().copied());
assert!(bv.capacity() >= 10);
bv.shrink_to_fit();
assert!(bv.capacity() >= 3);

pub fn into_boxed_slice(self) -> Box<[T]>

Converts the bit-vector into [Box<[T]>].

Note that this will drop any excess capacity.

For the vec-to-box equivalent that produces a BitBox<C, T>, see [into_boxed_bitslice].

Examples

let bv = bitvec![1, 0, 1];

let slice = bv.into_boxed_slice();

Any excess capacity is removed:

let mut bv = BitVec::<Local, usize>::with_capacity(100);
bv.extend([true, false, true].iter().copied());

assert!(bv.capacity() >= 100);
let slice = bv.into_boxed_slice();
let boxed_bitslice = BitBox::<Local, usize>::from_boxed_slice(slice);
let bv = BitVec::from_boxed_bitslice(boxed_bitslice);
assert!(bv.capacity() >= 3);

[Box<[T]>]: https://doc.rust-lang.org/std/boxed/struct.Box.html [into_boxed_bitslice]: #method.into_boxed_bitslice

pub fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len bits and dropping the rest.

If len is greater than the vector’s current length, this has no effect.

The [drain] method can emulate truncate, but causes the excess bits to be returned instead of dropped.

Note that this method has no effect on the allocated capacity of the vector.

Examples

Truncating a five-bit vector to two bits:

let mut bv = bitvec![1, 0, 1, 0, 1];
bv.truncate(2);
assert_eq!(bv, bitvec![1, 0]);

No truncation occurs when len is greater than the vector’s current length:

let mut bv = bitvec![1; 5];
bv.truncate(10);
assert_eq!(bv, bitvec![1; 5]);

Truncating to zero is equivalent to calling the [clear] method.

let mut bv = bitvec![0; 5];
bv.truncate(0);
assert!(bv.is_empty());

pub fn as_slice(&self) -> &[T]

Extracts an element slice containing the entire vector.

Unlike BitSlice::as_slice, this will produce partial edge elements, as they are known to not be aliased by any other slice handles.

Examples

use std::io::{self, Write};
let buffer = bitvec![Local, u8; 1, 0, 1, 0, 1];
io::sink().write(buffer.as_slice()).unwrap();

pub fn as_mut_slice(&mut self) -> &mut [T]

Extracts a mutable slice of the entire vector.

Unlike BitSlice::as_mut_slice, this will produce partial edge elements, as they are known to not be aliased by any other slice handles.

Examples

use std::io::{self, Read};
let mut buffer = bitvec![Local, u8; 0; 24];
io::repeat(0xA5u8).read_exact(buffer.as_mut_slice()).unwrap();

pub unsafe fn set_len(&mut self, new_len: usize)

Forces the length of the vector to new_len.

This is a low-level operation that maintains none of the normal invariants of the type. Normally changing the length of a vector is done using one of the safe operations instead, such as [truncate], [resize], [extend], or [clear].

Safety

new_len must be less than or equal to [capacity()].

The underlying elements at old_len ..new_len must be initialized.

Examples

This method can be useful for situations in which the vector is serving as a buffer for other code, particularly over FFI.

let mut bv = BitVec::<Local, usize>::with_capacity(17);
assert!(bv.is_empty());
unsafe { bv.set_len(23) };
assert_eq!(bv.len(), 23);

This example executes correctly, because the allocator can only reserve even multiples of bytes, and so rounds up from the with_capacity argument.

pub fn swap_remove(&mut self, index: usize) -> bool

Removes a bit from the vector and returns it.

The removed bit is replaced by the last bit of the vector.

This does not preserve ordering, but is O(1).

Panics

Panics if index is out of bounds.

Examples

let mut bv = bitvec![1, 0, 1, 0, 1];

assert!(!bv.swap_remove(1));
assert_eq!(bv, bitvec![1, 1, 1, 0]);

assert!(bv.swap_remove(0));
assert_eq!(bv, bitvec![0, 1, 1]);

pub fn insert(&mut self, index: usize, value: bool)

Inserts a bit at position index within the vector, shifting all bits after it to the right.

Panics

Panics if index > len.

Examples

let mut bv = bitvec![1, 0, 1, 0, 1];
bv.insert(1, false);
assert_eq!(bv, bitvec![1, 0, 0, 1, 0, 1]);
bv.insert(4, true);
assert_eq!(bv, bitvec![1, 0, 0, 1, 1, 0, 1]);

pub fn remove(&mut self, index: usize) -> bool

Removes and returns the bit at position index within the vector, shifting all bits after it to the left.

Panics

Panics if index is out of bounds.

Examples

let mut bv = bitvec![1, 0, 1, 0, 1];
assert!(!bv.remove(1));
assert_eq!(bv, bitvec![1, 1, 0, 1]);

pub fn retain<F>(&mut self, pred: F) where
F: FnMut(usize, bool) -> bool,

Retains only the bits that pass the predicate.

This removes all bits b where f(e) returns false. This method operates in place and preserves the order of the retained bits. Because it is in-place, it operates in O(n²) time.

API Differences

The Vec::retain method takes a predicate function with signature (&T) -> bool, whereas this method’s predicate function has signature (usize, &T) -> bool. This difference is in place because BitSlice by definition has only one bit of information per slice item, and including the index allows the callback function to make more informed choices.

Examples

let mut bv = bitvec![0, 1, 0, 1, 0, 1];
bv.retain(|_, b| b);
assert_eq!(bv, bitvec![1, 1, 1]);

pub fn push(&mut self, value: bool)

Appends a bit to the back of the vector.

If the vector is at capacity, this may cause a reallocation.

Panics

This will panic if the push will cause the vector to allocate above BitPtr<T>::MAX_ELTS or machine capacity.

Examples

let mut bv: BitVec = BitVec::new();
assert!(bv.is_empty());
bv.push(true);
assert_eq!(bv.len(), 1);
assert!(bv[0]);

pub fn pop(&mut self) -> Option<bool>

Removes the last element from a vector and returns it, or None if it is empty.

Examples

let mut bv: BitVec = BitVec::new();
assert!(bv.is_empty());
bv.push(true);
assert_eq!(bv.len(), 1);
assert!(bv[0]);

assert!(bv.pop().unwrap());
assert!(bv.is_empty());
assert!(bv.pop().is_none());

pub fn append<D, U>(&mut self, other: &mut BitVec<D, U>) where
D: BitOrder,
U: BitStore,

Moves all the elements of other into self, leaving other empty.

Panics

Panics if the number of bits in the vector overflows BitPtr::<T>::MAX_ELTS.

Examples

let mut bv1 = bitvec![0; 10];
let mut bv2 = bitvec![1; 10];
bv1.append(&mut bv2);
assert_eq!(bv1.len(), 20);
assert!(bv1[10]);
assert!(bv2.is_empty());

pub fn drain<R>(&mut self, range: R) -> Drain<'_, O, T> where
R: RangeBounds<usize>,

Creates a draining iterator that removes the specified range from the vector and yields the removed bits.

Notes

The element range is removed even if the iterator is only partially consumed or not consumed at all.

It is unspecified how many bits are removed from the vector if the Drain value is leaked.

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the vector.

Examples

let mut bv = bitvec![0, 0, 1, 1, 1, 0, 0];
assert_eq!(bv.len(), 7);
for bit in bv.drain(2 .. 5) {
  assert!(bit);
}
assert!(bv.not_any());
assert_eq!(bv.len(), 4);

pub fn clear(&mut self)

Clears the vector, removing all values.

Note that this method has no effect on the allocated capacity of the vector.

Examples

let mut bv = bitvec![1; 30];
assert_eq!(bv.len(), 30);
assert!(bv.iter().all(|b| *b));
bv.clear();
assert!(bv.is_empty());

After calling clear(), bv will no longer show raw memory, so the above test cannot show that the underlying memory is not altered. This is also an implementation detail on which you should not rely.

pub fn split_off(&mut self, at: usize) -> Self

Splits the collection into two at the given index.

Returns a newly allocated Self. self contains elements [0, at), and the returned Self contains elements [at, len).

Note that the capacity of self does not change.

Panics

Panics if at > len.

Examples

let mut bv1 = bitvec![0, 0, 0, 1, 1, 1];
let bv2 = bv1.split_off(3);
assert_eq!(bv1, bitvec![0, 0, 0]);
assert_eq!(bv2, bitvec![1, 1, 1]);

pub fn resize_with<F>(&mut self, new_len: usize, f: F) where
F: FnMut() -> bool,

Resizes the BitVec in-place so that len is equal to new_len.

If new_len is greater than len, the BitVec is extended by the difference, with each additional slot filled with the result of calling the closure f. The return values from f will end up in the BitVec in the order they have been generated.

If new_len is less than len, the BitVec is simply truncated.

This method uses a closure to create new values on every push. If you’d rather Clone a given value, use resize. If you want to use the Default trait to generate values, you can pass Default::default() as the second argument.

Examples

let mut bv = bitvec![1, 0, 1];
bv.resize_with(5, Default::default);
assert_eq!(bv, bitvec![1, 0, 1, 0, 0]);

let mut bv = bitvec![];
let mut p = 1;
bv.resize_with(4, || { p += 1; p % 2 == 0});
assert_eq!(bv, bitvec![1, 0, 1, 0]);

pub fn resize(&mut self, new_len: usize, value: bool)

Resizes the BitVec in place so that len is equal to new_len.

If new_len is greater than len, the BitVec is extended by the difference, with each additional slot filled with value. If new_len is less than len, the BitVec is simply truncated.

Examples

let mut bv = bitvec![0; 4];
bv.resize(8, true);
assert_eq!(bv, bitvec![0, 0, 0, 0, 1, 1, 1, 1]);
bv.resize(5, false);
assert_eq!(bv, bitvec![0, 0, 0, 0, 1]);

pub fn extend_from_slice<D, U>(&mut self, other: &BitSlice<D, U>) where
D: BitOrder,
U: BitStore,

Clones and appends all bits in a bit-slice to the BitVec.

Iterates over the bit-slice other, clones each bit, and then appends it to this BitVec. The other slice is traversed in-order.

Note that this function is the same as extend except that it is specialized to work with bit-slices instead. If and when Rust gets specialization this function will likely be deprecated (but still available).

Examples

let mut bv = bitvec![1];
bv.extend_from_slice(0xA5u8.bits::<Lsb0>());
assert_eq!(bv, bitvec![1, 1, 0, 1, 0, 0, 1, 0, 1]);

pub fn splice<R, I>(
&mut self,
range: R,
replace_with: I
) -> Splice<'_, O, T, <I as IntoIterator>::IntoIter> where
I: IntoIterator<Item = bool>,
R: RangeBounds<usize>,

Creates a splicing iterator that replaces the specified range in the vector with the given replace_with iterator and yields the removed bits. replace_with does not need to be the same length as range.

Notes

The element range is removed and replaced even if the iterator produced by this method is not consumed until the end.

It is unspecified how many bits are removed from the vector if the Splice value is leaked.

The input iterator replace_with is only consumed when the Splice value is dropped.

This is optimal if:

the tail (elements in the vector after range) is empty,

or replace_with yields fewer bits than range’s length,

the lower bound of its size_hint() is exact.

Otherwise, a temporary vector is allocated and the tail is moved twice.

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the vector.

Examples

This example starts with six bits of zero, and then splices out bits 2 and 3 and replaces them with four bits of one.

let mut bv = bitvec![0; 6];
let bv2 = bitvec![1; 4];

let s = bv.splice(2 .. 4, bv2).collect::<BitVec>();
assert_eq!(s.len(), 2);
assert!(!s[0]);
assert_eq!(bv, bitvec![0, 0, 1, 1, 1, 1, 0, 0]);

impl<O, T> BitVec<O, T> where
O: BitOrder,
T: BitStore,

pub fn repeat(bit: bool, len: usize) -> Self

Constructs a BitVec from a value repeated many times.

This function is equivalent to the bitvec![O, T; bit; len] macro call, and is in fact the implementation of that macro syntax.

Parameters

bit: The bit value to which all len allocated bits will be set.

len: The number of live bits in the constructed BitVec.

Returns

A BitVec with len live bits, all set to bit.

pub fn from_element(elt: T) -> Self

Constructs a BitVec from a single element.

The produced BitVec will span the element, and include all bits in it.

Parameters

elt: The source element.

Returns

A BitVec over the provided element.

Examples

use bitvec::prelude::*;

let bv = BitVec::<Msb0, u8>::from_element(5);
assert_eq!(bv.count_ones(), 2);

pub fn from_slice(slice: &[T]) -> Self

Constructs a BitVec from a slice of elements.

The produced BitVec will span the provided slice.

Parameters

slice: The source elements to copy into the new BitVec.

Returns

A BitVec set to the provided slice values.

Examples

use bitvec::prelude::*;

let src = [5, 10];
let bv = BitVec::<Msb0, u8>::from_slice(&src[..]);
assert!(bv[5]);
assert!(bv[7]);
assert!(bv[12]);
assert!(bv[14]);

pub fn from_vec(vec: Vec<T>) -> Self

Consumes a Vec<T> and creates a BitVec<C, T> from it.

Parameters

vec: The source vector whose memory will be used.

Returns

A new BitVec using the vec Vec’s memory.

Panics

Panics if the source vector would cause the BitVec to overflow capacity.

Examples

use bitvec::prelude::*;

let bv = BitVec::<Msb0, u8>::from_vec(vec![1, 2, 4, 8]);
assert_eq!(
  "[00000001, 00000010, 00000100, 00001000]",
  &format!("{}", bv),
);

pub fn from_bitslice(slice: &BitSlice<O, T>) -> Self

Clones a &BitSlice into a BitVec.

This is the only method by which a BitVec can be created whose first live bit is not at the 0 position. This behavior, though unconventional for common uses of BitVec, allows for a more efficient clone of any BitSlice region without shifting each bit in the region down to fit the 0 starting position.

Misaligned BitVecs do not have any adverse effect on usage other than the in-memory representation.

Whenever a BitVec is emptied, its head index is always set to 0, and will begin from the aligned position on future refills.

The ::force_align method will shift the BitVec’s data to begin at the 0 index, if you require this property.

Parameters

slice: The source BitSlice region. This may have any head index, and its memory will be copied element-wise into the new buffer.

Returns

A BitVec containing the same bits as the source slice.

Examples

use bitvec::prelude::*;

let bs = [0u8, !0].bits::<Msb0>();
let bv = BitVec::from_bitslice(bs);
assert_eq!(bv.len(), 16);
assert!(bv.some());

pub fn from_boxed_bitslice(slice: BitBox<O, T>) -> Self

Converts a frozen BitBox allocation into a growable BitVec.

This does not copy or reallocate.

Parameters

slice: A BitBox to be thawed.

Returns

A growable collection over the original memory of the slice.

Examples

use bitvec::prelude::*;

let bv = BitVec::from_boxed_bitslice(bitbox![0, 1]);
assert_eq!(bv.len(), 2);
assert!(bv.some());

pub unsafe fn from_raw_parts(pointer: BitPtr<T>, capacity: usize) -> Self

Creates a new BitVec<C, T> directly from the raw parts of another.

Parameters

pointer: The BitPtr<T> to use.

capacity: The number of T elements allocated in that slab.

Returns

A BitVec over the given slab of memory.

Safety

This is highly unsafe, due to the number of invariants that aren’t checked:

pointer needs to have been previously allocated by some allocating type.

pointer’s T needs to have the same size and alignment as it was initially allocated.

pointer’s element count needs to be less than or equal to the original allocation capacity.

capacity needs to be the original allocation capacity for the vector. This is not the value produced by .capacity().

Violating these will cause problems, like corrupting the handle’s concept of memory, the allocator’s internal data structures, and the sanity of your program. It is absolutely not safe to construct a BitVec whose T differs from the type used for the initial allocation.

The ownership of pointer is effectively transferred to the BitVec<C, T> which may then deallocate, reallocate, or modify the contents of the referent slice at will. Ensure that nothing else uses the pointer after calling this function.

pub fn as_bitslice(&self) -> &BitSlice<O, T>

Produces a BitSlice containing the entire vector.

Equivalent to &s[..].

Parameters

&self

Returns

A BitSlice over the vector.

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 1, 0];
let bs = bv.as_bitslice();

pub fn as_mut_bitslice(&mut self) -> &mut BitSlice<O, T>

Produces a mutable BitSlice containing the entire vector.

Equivalent to &mut s[..].

Parameters

&mut self

Returns

A mutable BitSlice over the vector.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 1, 0];
let bs = bv.as_mut_bitslice();

pub fn set_elements(&mut self, element: T)

Sets the backing storage to the provided element.

This unconditionally sets each live element in the backing buffer to the provided value, without altering the BitVec length or capacity. It operates on the underlying Vec’s memory buffer directly, and ignores the BitVec’s cursors.

It is an implementation detail as to whether this affects the value of allocated, but not currently used, elements in the buffer. Behavior of this method on elements not visible through self.as_slice() is not specified.

Parameters

&mut self

element: The value to which each live element in the backing store will be set.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![Local, u8; 0; 10];
//  note: the second element is not required to be zero, only to have
//  bits `0` and `1` according to `Local` be `0`.
assert_eq!(bv.as_slice()[0], 0);
bv.set_elements(0xA5);
assert_eq!(bv.as_slice(), &[0xA5, 0xA5]);

pub fn add_reverse<I>(self, addend: I) -> Self where
I: IntoIterator<Item = bool>,

Performs “reverse” addition (left to right instead of right to left).

This addition traverses the addends from left to right, performing the addition at each index and writing the sum into self.

If addend expires before self does, addend is zero-extended and the carry propagates through the rest of self. If self expires before addend, then self is zero-extended and the carry propagates through the rest of addend, growing self until addend expires.

An infinite addend will cause unbounded memory growth until the vector overflows and panics.

Parameters

self

addend: impl IntoIterator<Item=bool>: A stream of bits to add into self, from left to right.

Returns

The sum vector of self and addend.

Examples

use bitvec::prelude::*;

let a = bitvec![0, 1, 0, 1];
let b = bitvec![0, 0, 1, 1];
let c = a.add_reverse(b);
assert_eq!(c, bitvec![0, 1, 1, 0, 1]);

pub fn add_assign_reverse<I>(&mut self, addend: I) where
I: IntoIterator<Item = bool>,

Performs “reverse” addition (left to right instead of right to left).

This addition traverses the addends from left to right, performing the addition at each index and writing the sum into self.

If addend expires before self does, addend is zero-extended and the carry propagates through the rest of self. If self expires before addend, then self is zero-extended and the carry propagates through the rest of addend, growing self until addend expires.

An infinite addend will cause unbounded memory growth until the vector overflows and panics.

Parameters

&mut self

addend: impl IntoIterator<Item=bool>: A stream of bits to add into self, from left to right.

Effects

self may grow as a result of the final carry-out bit being 1 and pushed onto the right end.

Examples

use bitvec::prelude::*;

let mut a = bitvec![0, 1, 0, 1];
let     b = bitvec![0, 0, 1, 1];
a.add_assign_reverse(b.iter().copied());
assert_eq!(a, bitvec![0, 1, 1, 0, 1]);

pub fn change_order<P>(self) -> BitVec<P, T> where
P: BitOrder,

Changes the order type on the vector handle, without changing its contents.

Parameters

self

Returns

An equivalent vector handle with a new order type. The contents of the backing storage are unchanged.

To reorder the bits in memory, drain this vector into a new handle with the desired order type.

pub fn into_boxed_bitslice(self) -> BitBox<O, T>

Degrades a BitVec to a BitBox, freezing its size.

Parameters

self

Returns

Itself, with its size frozen and ungrowable.

pub fn into_vec(self) -> Vec<T>

Degrades a BitVec to a standard Vec.

Parameters

self

Returns

The plain vector underlying the BitVec.

pub fn force_align(&mut self)

Ensures that the live region of the underlying memory begins at the 0 bit position.

Notes

This method is currently implemented as a linear traversal that moves each bit individually from its original index to its final position. This is O(n) in the bit length of the vector.

It is possible to create an optimized rotation behavior that only moves a few bits individually, then moves elements in a gallop. The speed difference is proportional to the width of the element type.

When this behavior is implemented, force_align will be rewritten to take advantage of it. For now, it remains slow.

Examples

use bitvec::prelude::*;

let src = &bits![Msb0, u8; 1, 0, 1, 1, 0, 1, 1, 0][1 .. 7];
assert_eq!(src.len(), 6);
let mut bv = src.to_owned();
assert_eq!(bv.len(), 6);
assert_eq!(bv.as_slice()[0], 0xB6);
bv.force_align();
assert_eq!(bv.as_slice()[0], 0x6E);