Struct bitvec::vec::BitVec

#[repr(C)]
pub struct BitVec<T = usize, O = Lsb0> where
    T: BitStore,
    O: BitOrder,  { /* private fields */ }

Bit-Precision Dynamic Array

This is an analogue to Vec<bool> that stores its data using a compaction scheme to ensure that each bool takes exactly one bit of memory. It is similar to the C++ type std::vector<bool>, but uses bitvec’s type parameter system to provide more detailed control over the in-memory representation.

This is always a heap allocation. If you know your sizes at compile-time, you may prefer to use BitArray instead, which is able to store its data as an immediate value rather than through an indirection.

Documentation Practices

BitVec exactly replicates the API of the standard-library Vec type, including inherent methods, trait implementations, and relationships with the BitSlice slice analogue.

Items that are either direct ports, or renamed variants, of standard-library APIs will have a ## Original section that links to their standard-library documentation. Items that map to standard-library APIs but have a different API signature will also have an ## API Differences section that describes what the difference is, why it exists, and how to transform your code to fit it. For example:

Original

Vec<T>

API Differences

As with all bitvec data structures, this takes two type parameters <T, O> that govern the bit-vector’s storage representation in the underlying memory, and does not take a type parameter to govern what data type it stores (always bool)

Suggested Uses

BitVec is able to act as a compacted usize => bool dictionary, and is useful for holding large collections of truthiness. For instance, you might replace a Vec<Option<T>> with a (BitVec, Vec<MaybeUninit<T>>) to cut down on the resident size of the discriminant.

Through the BitField trait, BitVec is also able to act as a transport buffer for data that can be marshalled as integers. Serializing data to a narrower compacted form, or deserializing data from that form, can be easily accomplished by viewing subsets of a bit-vector and storing integers into, or loading integers out of, that subset. As an example, transporting four ten-bit integers can be done in five bytes instead of eight like so:

use bitvec::prelude::*;

let mut bv = bitvec![u8, Msb0; 0; 40];
bv[0 .. 10].store::<u16>(0x3A8);
bv[10 .. 20].store::<u16>(0x2F9);
bv[20 .. 30].store::<u16>(0x154);
bv[30 .. 40].store::<u16>(0x06D);

If you wish to use bit-field memory representations as struct fields rather than a transport buffer, consider BitArray instead: that type keeps its data as an immediate, and is more likely to act like a C struct with bitfields.

Examples

BitVec has exactly the same API as Vec<bool>, and even extends it with some of Vec<T>’s behaviors. As a brief tour:

Push and Pop

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();
bv.push(false);
bv.push(true);

assert_eq!(bv.len(), 2);
assert_eq!(bv[0], false);

assert_eq!(bv.pop(), Some(true));
assert_eq!(bv.len(), 1);

Writing Into a Bit-Vector

The only Vec<bool> API that BitVec does not implement is IndexMut, because that is not yet possible. Instead, .get_mut() can produce a proxy reference, or .set() can take an index and a value to write.

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();
bv.push(false);

*bv.get_mut(0).unwrap() = true;
assert!(bv[0]);
bv.set(0, false);
assert!(!bv[0]);

Macro Construction

Like Vec, BitVec also has a macro constructor: bitvec! takes a sequence of bit expressions and encodes them at compile-time into a suitable buffer. At run-time, this buffer is copied into the heap as a BitVec with no extra cost beyond the allocation.

use bitvec::prelude::*;

let bv = bitvec![0; 10];
let bv = bitvec![0, 1, 0, 0, 1];
let bv = bitvec![u16, Msb0; 1; 20];

Borrowing as BitSlice

BitVec lends its buffer as a BitSlice, so you can freely give permission to view or modify the contained data without affecting the allocation:

use bitvec::prelude::*;

fn read_bitslice(bits: &BitSlice) {
  // …
}

let bv = bitvec![0; 30];
read_bitslice(&bv);
let bs: &BitSlice = &bv;

Other Notes

The default type parameters are <usize, Lsb0>. This is the most performant pair when operating on memory, but likely does not match your needs if you are using BitVec to represent a transport buffer. See the user guide for more details on how the type parameters govern memory representation.

Applications, or single-purpose libraries, built atop bitvec will likely want to create a type alias with specific type parameters for their usage. bitvec is fully generic over the ordering/storage types, but this generality is rarely useful for client crates to propagate. <usize, Lsb0> is fastest; <u8, Msb0> matches what most debugger views of memory will print, and the rest are documented in the guide.

Safety

Unlike the other data structures in this crate, BitVec is uniquely able to hold uninitialized memory and produce pointers into it. As described in the BitAccess documentation, this crate is categorically unable to operate on uninitialized memory in any way. In particular, you may not allocate a buffer using ::with_capacity(), then use .as_mut_bitptr() to create a pointer used to write into the uninitialized buffer.

You must always initialize the buffer contents of a BitVec before attempting to view its contents. You can accomplish this through safe APIs such as .push(), .extend(), or .reserve(). These are all guaranteed to safely initialize the memory elements underlying the BitVec buffer without incurring undefined behavior in their operation.

Implementations

impl<T, O> BitVec<T, O> where
T: BitStore,
O: BitOrder,

Port of the Vec<T> inherent API.

pub fn new() -> Self

Constructs a new, empty, bit-vector.

This does not allocate until bits are .push()ed into it, or space is explicitly .reserve()d.

Original

Vec::new

Examples

use bitvec::prelude::*;

let bv = BitVec::<u8, Msb0>::new();
assert!(bv.is_empty());

pub fn with_capacity(capacity: usize) -> Self

Allocates a new, empty, bit-vector with space for at least capacity bits before reallocating.

Original

Vec::with_capacity

Panics

This panics if the requested capacity is longer than what the bit-vector can represent. See BitSlice::MAX_BITS.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(128);

assert!(bv.is_empty());
assert!(bv.capacity() >= 128);

for i in 0 .. 128 {
  bv.push(i & 0xC0 == i);
}
assert_eq!(bv.len(), 128);
assert!(bv.capacity() >= 128);

bv.push(false);
assert_eq!(bv.len(), 129);
assert!(bv.capacity() >= 129);

pub unsafe fn from_raw_parts(
bitptr: BitPtr<Mut, T, O>,
length: usize,
capacity: usize
) -> Self

Constructs a bit-vector handle from its constituent fields.

Original

Vec::from_raw_parts

Safety

The only acceptable argument values for this function are those that were previously produced by calling .into_raw_parts(). Furthermore, you may only call this at most once on any set of arguments. Using the same arguments in more than one call to this function will result in a double- or use-after free error.

Attempting to conjure your own values and pass them into this function will break the allocator state.

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0, 0, 1];
let (bitptr, len, capa) = bv.into_raw_parts();
let bv2 = unsafe {
  BitVec::from_raw_parts(bitptr, len, capa)
};
assert_eq!(bv2, bits![0, 1, 0, 0, 1]);

pub fn into_raw_parts(self) -> (BitPtr<Mut, T, O>, usize, usize)

Decomposes a bit-vector into its constituent member fields.

This disarms the destructor. In order to prevent a memory leak, you must pass these exact values back into ::from_raw_parts().

Original

Vec::into_raw_parts

API Differences

This method is still unstable as of 1.54. It is provided here as a convenience, under the expectation that the standard-library method will stabilize as-is.

pub fn capacity(&self) -> usize

Gets the allocation capacity, measured in bits.

This counts how many total bits the bit-vector can store before it must perform a reällocation to acquire more memory.

If the capacity is not a multiple of 8, you should call .force_align().

Original

Vec::capacity

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0, 0, 1];

pub fn reserve(&mut self, additional: usize)

Ensures that the bit-vector has allocation capacity for at least additional more bits to be appended to it.

For convenience, this method guarantees that the underlying memory for self[.. self.len() + additional] is initialized, and may be safely accessed directly without requiring use of .push() or .extend() to initialize it.

Newly-allocated memory is always initialized to zero. It is still dead until the bit-vector is grown (by .push(), .extend(), or .set_len()), but direct access will not trigger UB.

Original

Vec::reserve

Panics

This panics if the new capacity exceeds the bit-vector’s maximum.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(80);
assert!(bv.capacity() >= 80);
bv.reserve(800);
assert!(bv.capacity() >= 800);

pub fn reserve_exact(&mut self, additional: usize)

Ensures that the bit-vector has allocation capacity for at least additional more bits to be appended to it.

This differs from .reserve() by requesting that the allocator provide the minimum capacity necessary, rather than a potentially larger amount that the allocator may find more convenient.

Remember that this is a request: the allocator provides what it provides, and you cannot rely on the new capacity to be exactly minimal. You should still prefer .reserve(), especially if you expect to append to the bit-vector in the future.

Original

Vec::reserve_exact

Panics

This panics if the new capacity exceeds the bit-vector’s maximum.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(80);
assert!(bv.capacity() >= 80);
bv.reserve_exact(800);
assert!(bv.capacity() >= 800);

pub fn shrink_to_fit(&mut self)

Releases excess capacity back to the allocator.

Like .reserve_exact(), this is a request to the allocator, not a command. The allocator may reclaim excess memory or may not.

Original

Vec::shrink_to_fit

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(1000);
bv.push(true);
bv.shrink_to_fit();

pub fn into_boxed_slice(self) -> BitBox<T, O>

👎 Deprecated:
prefer `.into_boxed_bitslice() instead

pub fn truncate(&mut self, new_len: usize)

Shortens the bit-vector, keeping the first new_len bits and discarding the rest.

If len is greater than the bit-vector’s current length, this has no effect.

The .drain() method can emulate .truncate(), except that it yields the excess bits rather than discarding them.

Note that this has no effect on the allocated capacity of the bit-vector, nor does it erase truncated memory. Bits in the allocated memory that are outside of the .as_bitslice() view are always considered to have initialized, but unspecified, values, and you cannot rely on them to be zero.

Original

Vec::truncate

Examples

Truncating a five-bit vector to two bits:

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
bv.truncate(2);
assert_eq!(bv.len(), 2);
assert!(bv.as_raw_slice()[0].count_ones() >= 2);

No truncation occurs when len is greater than the bit-vector’s current length:

pub fn as_slice(&self) -> &BitSlice<T, O>

👎 Deprecated:
use .as_bitslice() instead

pub fn as_mut_slice(&mut self) -> &mut BitSlice<T, O>

👎 Deprecated:
use .as_mut_bitslice() instead

pub fn as_ptr(&self) -> BitPtr<Const, T, O>

👎 Deprecated:
use .as_bitptr() instead

pub fn as_mut_ptr(&mut self) -> BitPtr<Mut, T, O>

👎 Deprecated:
use .as_mut_bitptr() instead

pub unsafe fn set_len(&mut self, new_len: usize)

Resizes a bit-vector to a new length.

Original

Vec::set_len

Safety

NOT ALL MEMORY IN THE ALLOCATION IS INITIALIZED!

Memory in a bit-vector’s allocation is only initialized when the bit-vector grows into it normally (through .push() or one of the various .extend*() methods). Setting the length to a value beyond what was previously initialized, but still within the allocation, is undefined behavior.

The caller is responsible for ensuring that all memory up to (but not including) the new length has already been initialized.

Panics

This panics if new_len exceeds the capacity as reported by .capacity().

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
unsafe {
  // The default storage type, `usize`, is at least 32 bits.
  bv.set_len(32);
}
assert_eq!(bv, bits![
  0, 1, 0, 0, 1, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0,
]);
//  `BitVec` guarantees that newly-initialized memory is zeroed.

pub fn swap_remove(&mut self, index: usize) -> bool

Takes a bit out of the bit-vector.

The empty slot is filled with the last bit in the bit-vector, rather than shunting index + 1 .. self.len() down by one.

Original

Vec::swap_remove

Panics

This panics if index is out of bounds (self.len() or greater).

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
assert!(!bv.swap_remove(2));
assert_eq!(bv, bits![0, 1, 1, 0]);

pub fn insert(&mut self, index: usize, value: bool)

Inserts a bit at a given position, shifting all bits after it one spot to the right.

index may be any value up to and including self.len(). If it is self.len(), it behaves equivalently to .push().

Original

Vec::insert

Panics

This panics if index is out of bounds (including self.len()).

pub fn remove(&mut self, index: usize) -> bool

Removes a bit at a given position, shifting all bits after it one spot to the left.

index may be any value up to, but not including, self.len().

Original

Vec::remove

Panics

This panics if index is out of bounds (excluding self.len()).

pub fn retain<F>(&mut self, func: F) where
F: FnMut(usize, &bool) -> bool,

Retains only the bits that the predicate allows.

Bits are deleted from the vector when the predicate function returns false. This function is linear in self.len().

Original

Vec::retain

API Differences

The predicate receives both the index of the bit as well as its value, in order to allow the predicate to have more than one bit of keep/discard information.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
bv.retain(|idx, _| idx % 2 == 0);
assert_eq!(bv, bits![0,    0,    1]);

pub fn push(&mut self, value: bool)

Appends a single bit to the vector.

Original

Vec::push

Panics

This panics if the push would cause the bit-vector to exceed its maximum capacity.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0];
bv.push(true);
assert_eq!(bv.as_bitslice(), bits![0, 0, 1]);

pub fn pop(&mut self) -> Option<bool>

Attempts to remove the trailing bit from the bit-vector.

This returns None if the bit-vector is empty.

Original

Vec::pop

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1];
assert!(bv.pop().unwrap());
assert!(!bv.pop().unwrap());
assert!(bv.pop().is_none());

pub fn append<T2, O2>(&mut self, other: &mut BitVec<T2, O2>) where
T2: BitStore,
O2: BitOrder,

Moves all the bits out of other into the back of self.

The other bit-vector is emptied after this occurs.

Original

Vec::append

API Differences

This permits other to have different type parameters than self, and does not require that it be literally Self.

Panics

This panics if self.len() + other.len() exceeds the maximum capacity of a bit-vector.

Examples

use bitvec::prelude::*;

let mut bv1 = bitvec![u16, Msb0; 0; 10];
let mut bv2 = bitvec![u32, Lsb0; 1; 10];

bv1.append(&mut bv2);

assert_eq!(bv1.count_ones(), 10);
assert_eq!(bv1.count_zeros(), 10);
assert!(bv2.is_empty());

pub fn drain<R>(&mut self, range: R) -> Drain<'_, T, O> where
R: RangeBounds<usize>,

Iterates over a portion of the bit-vector, removing all yielded bits from it.

When the iterator drops, all bits in its coverage are removed from self, even if the iterator did not yield them. If the iterator is leaked or otherwise forgotten, and its destructor never runs, then the amount of un-yielded bits removed from the bit-vector is not specified.

Original

Vec::drain

Panics

This panics if range departs 0 .. self.len().

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
let bv2 = bv.drain(1 ..= 3).collect::<BitVec>();
assert_eq!(bv, bits![0,          1]);
assert_eq!(bv2, bits![1, 0, 0]);

// A full range clears the bit-vector.
bv.drain(..);
assert!(bv.is_empty());

pub fn clear(&mut self)

Empties the bit-vector.

This does not affect the allocated capacity.

Original

Vec::clear

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
bv.clear();
assert!(bv.is_empty());

pub fn len(&self) -> usize

Gets the length of the bit-vector.

This is equivalent to BitSlice::len; it is provided as an inherent method here rather than relying on Deref forwarding so that you can write BitVec::len as a named function item.

Original

Vec::len

pub fn is_empty(&self) -> bool

Tests if the bit-vector is empty.

This is equivalent to BitSlice::is_empty; it is provided as an inherent method here rather than relying on Deref forwarding so that you can write BitVec::is_empty as a named function item.

Original

Vec::is_empty

pub fn split_off(&mut self, at: usize) -> Self

Splits the bit-vector in half at an index, moving self[at ..] out into a new bit-vector.

Original

Vec::split_off

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
let bv2 = bv.split_off(2);
assert_eq!((&*bv, &*bv2), (bits![0, 1], bits![0, 0, 1]));

pub fn resize_with<F>(&mut self, new_len: usize, func: F) where
F: FnMut(usize) -> bool,

Resizes the bit-vector to a new length, using a function to produce each inserted bit.

If new_len is less than self.len(), this is a truncate operation; if it is greater, then self is extended by repeatedly pushing func().

Original

Vec::resize_with

API Differences

The generator function receives the index into which its bit will be placed.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 2];
bv.resize_with(5, |idx| idx % 2 == 1);
assert_eq!(bv, bits![1, 1, 0, 1, 0]);

pub fn leak<'a>(self) -> &'a mut BitSlice<T, O>

Destroys the BitVec handle without destroying the bit-vector allocation. The allocation is returned as an &mut BitSlice that lasts for the remaining program lifetime.

You may call BitBox::from_raw on this slice handle exactly once in order to reap the allocation before program exit. That function takes a mutable pointer, not a mutable reference, so you must ensure that the returned reference is never used again after restoring the allocation handle.

Original

Vec::leak

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 0, 1];
let static_bits: &'static mut BitSlice = bv.leak();
static_bits.set(0, true);
assert_eq!(static_bits, bits![1, 0, 1]);

let bb = unsafe { BitBox::from_raw(static_bits) };
// static_bits may no longer be used.
drop(bb); // explicitly reap memory before program exit

pub fn resize(&mut self, new_len: usize, value: bool)

Resizes the bit-vector to a new length. New bits are initialized to value.

Original

Vec::resize

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0; 2];
bv.resize(5, true);
assert_eq!(bv, bits![0, 0, 1, 1, 1]);

pub fn extend_from_slice<T2, O2>(&mut self, other: &BitSlice<T2, O2>) where
T2: BitStore,
O2: BitOrder,

👎 Deprecated:
use .extend_from_bitslice() or .extend_from_raw_slice() instead

pub fn extend_from_within<R>(&mut self, src: R) where
R: RangeExt<usize>,

Extends self by copying an internal range of its bit-slice as the region to append.

Original

Vec::extend_from_within

Panics

This panics if src is not within 0 .. self.len().

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
bv.extend_from_within(1 .. 4);
assert_eq!(bv, bits![0, 1, 0, 0, 1, 1, 0, 0]);

pub fn splice<R, I>(
&mut self,
range: R,
replace_with: I
) -> Splice<'_, T, O, I::IntoIter> where
R: RangeBounds<usize>,
I: IntoIterator<Item = bool>,

Modifies self.drain() so that the removed bit-slice is instead replaced with the contents of another bit-stream.

As with .drain(), the specified range is always removed from the bit-vector even if the splicer is not fully consumed, and the splicer does not specify how many bits are removed if it leaks.

The replacement source is only consumed when the splicer drops; however, it may be pulled before then. The replacement source cannot assume that there will be a delay between creation of the splicer and when it must begin producing bits.

This copies the Vec::splice implementation; see its documentation for more details about how the replacement should act.

Original

Vec::splice

Panics

This panics if range departs 0 .. self.len().

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 1];
//                   a  b  c
let mut yank = bv.splice(
  .. 2,
  bits![static 1, 1, 0].iter().by_vals(),
//             d  e  f
);

assert!(!yank.next().unwrap()); // a
assert!(yank.next().unwrap()); // b
drop(yank);
assert_eq!(bv, bits![1, 1, 0, 1]);
//                   d  e  f  c

impl<T, O> BitVec<T, O> where
T: BitStore,
O: BitOrder,

Constructors.

pub const EMPTY: Self = Self { bitspan: BitSpan::EMPTY, capacity: 0, }

An empty bit-vector with no backing allocation.

pub fn repeat(bit: bool, len: usize) -> Self

Creates a new bit-vector by repeating a bit for the desired length.

Examples

use bitvec::prelude::*;

let zeros = BitVec::<u8, Msb0>::repeat(false, 50);
let ones = BitVec::<u16, Lsb0>::repeat(true, 50);

pub fn from_bitslice(slice: &BitSlice<T, O>) -> Self

Copies the contents of a bit-slice into a new heap allocation.

This copies the raw underlying elements into a new allocation, and sets the produced bit-vector to use the same memory layout as the originating bit-slice. This means that it may begin at any bit in the first element, not just the zeroth bit. If you require this property, call .force_align().

Dead bits in the copied memory elements are guaranteed to be zeroed.

Examples

use bitvec::prelude::*;

let bits = bits![0, 1, 0, 0, 1];
let bv = BitVec::from_bitslice(bits);
assert_eq!(bv, bits);

pub fn from_element(elem: T) -> Self

Constructs a new bit-vector from a single element.

This copies elem into a new heap allocation, and sets the bit-vector to cover it entirely.

Examples

use bitvec::prelude::*;

let bv = BitVec::<_, Msb0>::from_element(1u8);
assert!(bv[7]);

pub fn from_slice(slice: &[T]) -> Self

Constructs a new bit-vector from a slice of memory elements.

This copies slice into a new heap allocation, and sets the bit-vector to cover it entirely.

Panics

This panics if slice exceeds bit-vector capacity.

Examples

use bitvec::prelude::*;

let slice = &[0u8, 1, 2, 3];
let bv = BitVec::<_, Lsb0>::from_slice(slice);
assert_eq!(bv.len(), 32);

pub fn try_from_slice(slice: &[T]) -> Result<Self, BitSpanError<T>>

Fallibly constructs a new bit-vector from a slice of memory elements.

This fails early if slice exceeds bit-vector capacity. If it is not, then slice is copied into a new heap allocation and fully spanned by the returned bit-vector.

Examples

use bitvec::prelude::*;

let slice = &[0u8, 1, 2, 3];
let bv = BitVec::<_, Lsb0>::try_from_slice(slice).unwrap();
assert_eq!(bv.len(), 32);

pub fn from_vec(vec: Vec<T>) -> Self

Converts a regular vector in-place into a bit-vector.

The produced bit-vector spans every bit in the original vector. No reällocation occurs; this is purely a transform of the handle.

Panics

This panics if the source vector is too long to view as a bit-slice.

Examples

use bitvec::prelude::*;

let v = vec![0u8, 1, 2, 3];
let bv = BitVec::<_, Msb0>::from_vec(v);
assert_eq!(bv.len(), 32);

pub fn try_from_vec(vec: Vec<T>) -> Result<Self, Vec<T>>

Attempts to convert a regular vector in-place into a bit-vector.

This fails if the source vector is too long to view as a bit-slice. On success, the produced bit-vector spans every bit in the original vector. No reällocation occurs; this is purely a transform of the handle.

Examples

use bitvec::prelude::*;

let v = vec![0u8; 20];
assert_eq!(BitVec::<_, Msb0>::try_from_vec(v).unwrap().len(), 160);

It is not practical to allocate a vector that will fail this conversion.

pub fn extend_from_bitslice<T2, O2>(&mut self, other: &BitSlice<T2, O2>) where
T2: BitStore,
O2: BitOrder,

Appends the contents of a bit-slice to a bit-vector.

This can extend from a bit-slice of any type parameters; it is not restricted to using the same parameters as self. However, when the type parameters do match, it is possible for this to use a batch-copy optimization to go faster than the individual-bit crawl that is necessary when they differ.

Until Rust provides extensive support for specialization in trait implementations, you should use this method whenever you are extending from a BitSlice proper, and only use the general .extend() implementation if you are required to use a generic bool source.

Original

Vec::extend_from_slice

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1];
bv.extend_from_bitslice(bits![0, 1, 0, 0, 1]);
assert_eq!(bv, bits![0, 1, 0, 1, 0, 0, 1]);

pub fn extend_from_raw_slice(&mut self, slice: &[T])

Appends a slice of T elements to a bit-vector.

The slice is viewed as a BitSlice<T, O>, then appended directly to the bit-vector.

Original

Vec::extend_from_slice

impl<T, O> BitVec<T, O> where
T: BitStore,
O: BitOrder,

Converters.

pub fn as_bitslice(&self) -> &BitSlice<T, O>

Explicitly views the bit-vector as a bit-slice.

pub fn as_mut_bitslice(&mut self) -> &mut BitSlice<T, O>

Explicitly views the bit-vector as a mutable bit-slice.

pub fn as_raw_slice(&self) -> &[T]

Views the bit-vector as a slice of its underlying memory elements.

pub fn as_raw_mut_slice(&mut self) -> &mut [T]

Views the bit-vector as a mutable slice of its underlying memory elements.

pub fn as_bitptr(&self) -> BitPtr<Const, T, O>

Creates an unsafe shared bit-pointer to the start of the buffer.

Original

Vec::as_ptr

Safety

You must initialize the contents of the underlying buffer before accessing memory through this pointer. See the BitPtr documentation for more details.

pub fn as_mut_bitptr(&mut self) -> BitPtr<Mut, T, O>

Creates an unsafe writable bit-pointer to the start of the buffer.

Original

Vec::as_mut_ptr

Safety

You must initialize the contents of the underlying buffer before accessing memory through this pointer. See the BitPtr documentation for more details.

pub fn into_boxed_bitslice(self) -> BitBox<T, O>

Converts a bit-vector into a boxed bit-slice.

This may cause a reällocation to drop any excess capacity.

Original

Vec::into_boxed_slice

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0, 0, 1];
let bb = bv.into_boxed_bitslice();

pub fn into_vec(self) -> Vec<T>

Converts a bit-vector into a Vec of its underlying storage.

The produced vector contains all elements that contained live bits. Dead bits have an unspecified value; you should call .set_uninitialized() before converting into a vector.

This does not affect the allocated memory; it is purely a conversion of the handle.

Examples

use bitvec::prelude::*;

let bv = bitvec![u8, Msb0; 0, 1, 0, 0, 1];
let v = bv.into_vec();
assert_eq!(v[0] & 0xF8, 0b01001_000);

impl<T, O> BitVec<T, O> where
T: BitStore,
O: BitOrder,

Utilities.

pub fn set_elements(&mut self, element: T::Mem)

Overwrites each element (visible in .as_raw_mut_slice()) with a new bit-pattern.

This unconditionally writes element into each element in the backing slice, without altering the bit-vector’s length or capacity.

This guarantees that dead bits visible in .as_raw_slice() but not .as_bitslice() are initialized according to the bit-pattern of element. The elements not visible in the raw slice, but present in the allocation, do not specify a value. You may not rely on them being zeroed or being set to the element bit-pattern.

Parameters

&mut self

element: The bit-pattern with which each live element in the backing store is initialized.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![u8, Msb0; 0; 20];
assert_eq!(bv.as_raw_slice(), [0; 3]);
bv.set_elements(0xA5);
assert_eq!(bv.as_raw_slice(), [0xA5; 3]);

pub fn set_uninitialized(&mut self, value: bool)

Sets the uninitialized bits of a bit-vector to a known value.

This method modifies all bits that are observable in .as_raw_slice() but not observable in .as_bitslice() to a known value. Memory beyond the raw-slice view, but still within the allocation, is considered fully dead and will never be seen.

This can be used to zero the unused memory so that when viewed as a raw slice, unused bits have a consistent and predictable value.

Examples

use bitvec::prelude::*;

let mut bv = 0b1101_1100u8.view_bits::<Lsb0>().to_bitvec();
assert_eq!(bv.as_raw_slice()[0], 0b1101_1100u8);

bv.truncate(4);
assert_eq!(bv.count_ones(), 2);
assert_eq!(bv.as_raw_slice()[0], 0b1101_1100u8);

bv.set_uninitialized(false);
assert_eq!(bv.as_raw_slice()[0], 0b0000_1100u8);

bv.set_uninitialized(true);
assert_eq!(bv.as_raw_slice()[0], 0b1111_1100u8);

pub fn force_align(&mut self)

Ensures that the live region of the bit-vector’s contents begin at the front edge of the buffer.

BitVec has performance optimizations where it moves its view of its buffer contents in order to avoid needless moves of its data within the buffer. This can lead to unexpected contents of the raw memory values, so this method ensures that the semantic contents of the bit-vector match its in-memory storage.

Examples

use bitvec::prelude::*;

let data = 0b00_1111_00u8;
let bits = data.view_bits::<Msb0>();

let mut bv = bits[2 .. 6].to_bitvec();
assert_eq!(bv, bits![1; 4]);
assert_eq!(bv.as_raw_slice()[0], data);

bv.force_align();
assert_eq!(bv, bits![1; 4]);
// BitVec does not specify the value of dead bits in its buffer.
assert_eq!(bv.as_raw_slice()[0] & 0xF0, 0xF0);