Struct bitvec::vec::BitVec

#[repr(C)]
pub struct BitVec<O = Lsb0, T = usize> where
    O: BitOrder,
    T: BitStore,  { /* private fields */ }

A contiguous growable array of bits.

This is a managed, heap-allocated, buffer that contains a BitSlice region. It is analagous to Vec<bool>, and is written to be very nearly a drop-in replacement for it. This type contains little interesting behavior in its own right; most of its behavior is provided by dereferencing to its managed BitSlice buffer. It instead serves primarily as an interface to the allocator, and has some specialized behaviors for its fully-owned memory buffer.

Documentation

All APIs that mirror something in the standard library will have an Original section linking to the corresponding item. All APIs that have a different signature or behavior than the original will have an API Differences section explaining what has changed, and how to adapt your existing code to the change.

These sections look like this:

Original

Vec<T>

API Differences

The buffer type Vec<bool> has no type parameters. BitVec<O, T> has the same two type parameters as BitSlice<O, T>. Otherwise, BitVec is able to implement the full API surface of Vec<bool>.

Examples

Because BitVec takes type parameters, but has default type arguments for them, you will need to specify its type parameters when using its associated functions. The easiest way to do this is to declare bindings type as : BitVec, which uses the default type arguments.

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();
bv.push(false);
bv.push(true);

assert_eq!(bv.len(), 2);
assert_eq!(bv[0], false);

assert_eq!(bv.pop(), Some(true));
assert_eq!(bv.len(), 1);

// `BitVec` cannot yet support `[]=` write indexing.
*bv.get_mut(0).unwrap() = true;
assert_eq!(bv[0], true);

bv.extend(bits![0, 1, 0]);

for bit in &bv {
  println!("{}", bit);
}
assert_eq!(bv, bits![1, 0, 1, 0]);

The bitvec! macro is provided to make initialization more convenient:

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1];
bv.push(true);
assert_eq!(bv, bits![0, 0, 1, 1]);

It has the same argument syntax as vec!. In addition, it can take type arguments for ordering and storage:

use bitvec::prelude::*;

let bv = bitvec![Msb0, u16; 1; 30];
assert!(bv.all());
assert_eq!(bv.len(), 30);

Indexing

The BitVec type allows you to access bits by index, because it implements the Index trait. However, because IndexMut requires producing an &mut bool reference, it cannot implement []= index assignment syntax. Instead, you must use get_mut or get_unchecked_mut to produce proxy types that can serve the same purpose.

Slicing

A BitVec is resizable, while BitSlice is a fixed-size view of a buffer. Just as with ordinary Vecs and slices, you can get a BitSlice from a BitVec by borrowing it:

use bitvec::prelude::*;

fn read_bitslice(slice: &BitSlice) {
  // …
}

let bv = bitvec![0; 30];
read_bitslice(&bv);

// … and that’s all!
// you can also do it like this:
let x: &BitSlice = &bv;

As with ordinary Rust types, you should prefer passing bit-slices rather than buffers when you just want to inspect the data, and not manage the underlying memory region.

Behavior

Because BitVec is a fully-owned buffer, it is able to operate on its memory without concern for any other views that may alias. This enables it to specialize some BitSlice behavior to be faster or more efficient. However, BitVec is not restricted to only using unaliased integer storage, and technically permits the construction of BitVec<_, AtomicType>.

This restriction is extremely awkward and constraining to write in the library, and clients will probably never attempt to construct them, but the possibility is still present. Be aware of this possibility when using generic code to convert from BitSlice to BitVec. Fully-typed code does not need to be concerned with this possibility.

Capacity and Reällocation

The capacity of a bit-vector is the amount of space allocated for any future bits that will be added onto the vector. This is not to be confused with the length of a vector, which specifies the number of actual bits within the vector. If a vector’s length exceeds its capacity, its capacity will automatically be increased, but its buffer will have to be reällocated

For example, a bit-vector with capacity 64 and length 0 would be an empty vector with space for 64 more bits. Pushing 64 or fewer bits onto the vector will not change its capacity or cause reällocation to occur. However, if the vector’s length is increased to 65, it may have to reällocate, which can be slow. For this reason, it is recommended to use BitVec::with_capacity whenever possible to specify how big the vector is expected to get.

Safety

Like BitSlice, BitVec is exactly equal in size to Vec, and is also absolutely representation-incompatible with it. You must never attempt to type-cast between Vec<T> and BitVec in any way, nor attempt to modify the memory value of a BitVec handle. Doing so will cause allocator and memory errors in your program, likely inducing a panic.

Everything in the BitVec public API, even the unsafe parts, are guaranteed to have no more unsafety than their equivalent items in the standard library. All unsafe APIs will have documentation explicitly detailing what the API requires you to uphold in order for it to function safely and correctly. All safe APIs will do so themselves.

Performance

The choice of BitStore type parameter can impact your vector’s performance, as the allocator operates in units of T rather than in bits. This means that larger register types will increase the amount of memory reserved in each call to the allocator, meaning fewer calls to push will actually cause a reällocation. In addition, iteration over the vector is governed by the BitSlice characteristics on the type parameter. You are generally better off using larger types when your vector is a data collection rather than a specific I/O protocol buffer.

Macro Construction

Heap allocation can only occur at runtime, but the bitvec! macro will construct an appropriate BitSlice buffer at compile-time, and at run-time, only copy the buffer into a heap allocation.

Implementations

impl<O, T> BitVec<O, T> where
O: BitOrder,
T: BitStore,

Port of the Vec<T> inherent API.

pub fn new() -> Self

Constructs a new, empty, BitVec<O, T>.

The bit-vector will not allocate until bits are pushed onto it.

Original

Vec::new

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty, BitVec<O, T> with the specified capacity (in bits).

The bit-vector will be able to hold at least capacity bits without reällocating. If capacity is 0, the bit-vector will not allocate.

It is important to note that although the returned bit-vector has the capacity specified, the bit-vector will have a zero length. For an explanation of the difference between length and capacity, see Capacity and reällocation.

Original

Vec::with_capacity

Panics

Panics if the requested capacity exceeds the bit-vector’s limits.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(128);

// The bit-vector contains no bits, even
// though it has the capacity for more.
assert_eq!(bv.len(), 0);
assert!(bv.capacity() >= 128);

// These are all done
// without reällocating…
for i in 0 .. 128 {
  bv.push(i & 0xC0 == i);
}
assert_eq!(bv.len(), 128);
assert!(bv.capacity() >= 128);

// …but this may make the
// bit-vector reällocate.
bv.push(false);
assert_eq!(bv.len(), 129);
assert!(bv.capacity() >= 129);

pub fn into_raw_parts(self) -> (BitPtr<Mut, O, T>, usize, usize)

Decomposes a BitVec<O, T> into its raw components.

Returns the raw bit-pointer to the underlying data, the length of the bit-vector (in bits), and the allocated capacity of the buffer (in bits). These are the same arguments in the same order as the arguments to from_raw_parts.

After calling this function, the caller is responsible for the memory previously managed by the BitVec. The only way to do this is to convert the raw bit-pointer, length, and capacity back into a BitVec with the from_raw_parts function, allowing the destructor to perform the cleanup.

Original

Vec::into_raw_parts

API Differences

This returns a BitPtr, rather than a *mut T. If you need the actual memory address, BitPtr::pointer will produce it.

Examples

use bitvec::prelude::*;
use core::cell::Cell;

let bv: BitVec = bitvec![0, 1, 0, 0, 1];

let (ptr, len, cap) = bv.into_raw_parts();

let rebuilt = unsafe {
  // We can now make changes to the components, such
  // as casting the pointer to a compatible type.
  let ptr = ptr.cast::<Cell<usize>>();
  BitVec::from_raw_parts(ptr, len, cap)
};
assert_eq!(rebuilt, bits![0, 1, 0, 0, 1]);

pub unsafe fn from_raw_parts(
bitptr: BitPtr<Mut, O, T>,
length: usize,
capacity: usize
) -> Self

Creates a BitVec<O, T> directly from the raw components of another bit-vector.

Original

Vec::from_raw_parts

API Differences

This takes a BitPtr, rather than a *mut T. If you only have a pointer, you can construct a BitPtr to its zeroth bit before calling this.

Safety

This is highly unsafe, due to the number of invariants that aren’t checked:

bitptr needs to have been previously allocated by BitVec<O, T>, or constructed from a pointer allocated by Vec<T>.

T needs to have the same size and alignment as what bitptr was allocated with. (T having a less strict alignment is not sufficient; the alignment really needs to be equal to satisfy the dealloc requirement that memory must be allocated and deällocated with the same layout.) However, you can safely cast between bare integers, BitSafe integers, Cell wrappers, and atomic integers, as long as they all have the same width.

length needs to be less than or equal to capacity.

capacity needs to be the capacity (in bits) that the bit-pointer was allocated with (less any head offset in bitptr).

Violating these will cause problems. For example, it is not safe to build a BitVec<_, u8> from a pointer to a u16 sequence and twice its original length, because the allocator cares about the alignment, and these two types have different alignments. The buffer was allocated with alignment 2 (for u16), but after turning it into a BitVec<_, u8>, it’ll be deällocated with alignment 1.

The ownership of bitptris effectively transferred to the BitVec<O, T> which may then deällocate, reällocate, or change the contents of memory pointed to by the bit-pointer at will. Ensure that nothing else uses the pointer after calling this function.

Examples

use bitvec::prelude::*;
use bitvec::ptr as bv_ptr;
use core::mem::ManuallyDrop;

let bv = bitvec![0, 1, 0, 0, 1];
let mut bv = ManuallyDrop::new(bv);
let bp = bv.as_mut_bitptr();
let len = bv.len();
let cap = bv.capacity();

unsafe {
  // Overwrite memory with the inverse bits.
  for i in 0 .. len {
    let bp = bp.add(i);
    bv_ptr::write(bp, !bv_ptr::read(bp.immut()));
  }

  // Put everything back together into a `BitVec`.
  let rebuilt = BitVec::from_raw_parts(bp, len, cap);
  assert_eq!(rebuilt, bits![1, 0, 1, 1, 0]);
}

pub fn capacity(&self) -> usize

Returns the number of bits the bit-vector can hold without reällocating.

Original

Vec::capacity

Examples

use bitvec::prelude::*;

let bv: BitVec = BitVec::with_capacity(100);
assert!(bv.capacity() >= 100);

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more bits to be inserted in the given BitVec<O, T>. The collection may reserve more space to avoid frequent reällocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

Original

Vec::reserve

Panics

Panics if the new capacity exceeds the bit-vector’s limits.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1];
bv.reserve(100);
assert!(bv.capacity() >= 101);

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more bits to be inserted in the given BitVec<O, T>. After calling reserve_exact, capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore, capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Original

Vec::reserve_exact

Panics

Panics if the new capacity exceeds the vector’s limits.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1];
bv.reserve_exact(100);
assert!(bv.capacity() >= 101);

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the bit-vector as much as possible.

It will drop down as close as possible to the length, but the allocator may still inform the bit-vector that there is space for a few more bits.

Original

Vec::shrink_to_fit

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(100);
bv.extend([false, true, false].iter().copied());
assert!(bv.capacity() >= 100);
bv.shrink_to_fit();
assert!(bv.capacity() >= 3);

pub fn truncate(&mut self, len: usize)

Shortens the bit-vector, keeping the first len bits and dropping the rest.

If len is greater than the bit-vector’s current length, this has no effect.

The drain method can emulate truncate, but causes the excess bits to be returned instead of dropped.

Note that this method has no effect on the allocated capacity of the bit-vector, nor does it erase truncated memory. Bits in the allocated memory that are outside of the as_bitslice view always have unspecified values, and cannot be relied upon to be zero.

Original

Vec::truncate

Examples

Truncating a five-bit vector to two bits:

use bitvec::prelude::*;

let mut bv = bitvec![1; 5];
bv.truncate(2);
assert_eq!(bv.len(), 2);
assert!(bv.as_raw_slice()[0].count_ones() >= 5);

No truncation occurs when len is greater than the vector’s current length:

use bitvec::prelude::*;

let mut bv = bitvec![1; 3];
bv.truncate(8);
assert_eq!(bv.len(), 3);

Truncating when len == 0 is equivalent to calling the clear method.

use bitvec::prelude::*;

let mut bv = bitvec![0; 3];
bv.truncate(0);
assert!(bv.is_empty());

pub unsafe fn set_len(&mut self, new_len: usize)

Forces the length of the bit-vector to new_len.

This is a low-level operation that maintains none of the normal invariants of the type. Normally changing the length of a bit-vector is done using one of the safe operations instead, such as truncate, resize, extend, or clear.

Original

Vec::set_len

Safety

new_len must be less than or equal to self.capacity().

The memory elements underlying old_len .. new_len must be initialized.

Examples

This method can be useful for situations in which the bit-vector is serving as a buffer for other code, particularly over FFI:

use bitvec::prelude::*;

// `bitvec` could pair with `rustler` for a better bitstream
type ErlBitstring = BitVec<Msb0, u8>;
let mut bits_read = 0;
// An imaginary Erlang function wants a large bit buffer.
let mut buf = ErlBitstring::with_capacity(32_768);
// SAFETY: When `erl_read_bits` returns `ERL_OK`, it holds that:
// 1. `bits_read` bits were initialized.
// 2. `bits_read` <= the capacity (32_768)
// which makes `set_len` safe to call.
unsafe {
  // Make the FFI call…
  let status = erl_read_bits(&mut buf, 10, &mut bits_read);
  if status == ERL_OK {
    // …and update the length to what was read in.
    buf.set_len(bits_read);
  }
}

pub fn swap_remove(&mut self, index: usize) -> bool

Removes a bit from the bit-vector and returns it.

The removed bit is replaced by the last bit of the bit-vector.

This does not preserve ordering, but is O(1).

Original

Vec::swap_remove

Panics

Panics if index is out of bounds.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1, 0, 1];
assert!(!bv.swap_remove(1));
assert_eq!(bv, bits![0, 1, 1, 0]);

assert!(!bv.swap_remove(0));
assert_eq!(bv, bits![0, 1, 1]);

pub fn insert(&mut self, index: usize, value: bool)

Inserts a bit at position index within the bit-vector, shifting all bits after it to the right.

Original

Vec::insert

Panics

Panics if index > len.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0; 5];
bv.insert(5, true);
assert_eq!(bv, bits![0, 0, 0, 0, 0, 1]);
bv.insert(2, true);
assert_eq!(bv, bits![0, 0, 1, 0, 0, 0, 1]);

pub fn remove(&mut self, index: usize) -> bool

Removes and returns the bit at position index within the bit-vector, shifting all bits after it to the left.

Original

Vec::remove

Panics

Panics if index is out of bounds.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0];
assert!(bv.remove(1));
assert_eq!(bv, bits![0, 0]);

pub fn retain<F>(&mut self, func: F) where
F: FnMut(usize, &bool) -> bool,

Retains only the bits specified by the predicate.

In other words, remove all bits b such that func(idx(b), &b) returns false. This method operates in place, visiting each bit exactly once in the original order, and preserves the order of the retained bits.

Original

Vec::retain

API Differences

In order to allow more than one bit of information for the retention decision, the predicate receives the index of each bit, as well as its value.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 1, 0, 0, 1];
bv.retain(|i, b| (i % 2 == 0) ^ b);
assert_eq!(bv, bits![0, 1, 0, 1]);

pub fn push(&mut self, value: bool)

Appends a bit to the back of a collection.

Original

Vec::push

Panics

Panics if the number of bits in the bit-vector exceeds the maximum bit-vector capacity.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0];
bv.push(true);
assert_eq!(bv.count_ones(), 1);

pub fn pop(&mut self) -> Option<bool>

Removes the last bit from a bit-vector and returns it, or None if it is empty.

Original

Vec::pop

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1];
assert_eq!(bv.pop(), Some(true));
assert_eq!(bv, bits![0, 0]);

pub fn append<O2, T2>(&mut self, other: &mut BitVec<O2, T2>) where
O2: BitOrder,
T2: BitStore,

Moves all the bits of other into self, leaving other empty.

Original

Vec::append

API Differences

This permits other to have different type parameters than self, and does not require that it be of literally Self.

Panics

Panics if the number of bits overflows the maximum bit-vector capacity.

Examples

use bitvec::prelude::*;

let mut bv1 = bitvec![Msb0, u16; 0; 10];
let mut bv2 = bitvec![Lsb0, u32; 1; 10];

bv1.append(&mut bv2);

assert_eq!(bv1.count_ones(), 10);
assert!(bv2.is_empty());

pub fn drain<R>(&mut self, range: R) -> Drain<'_, O, T> where
R: RangeBounds<usize>,

Creates a draining iterator that removes the specified range in the bit-vector and yields the removed bits.

When the iterator is dropped, all bits in the range are removed from the bit-vector, even if the iterator was not fully consumed. If the iterator is not dropped (with mem::forget for example), it is unspecified how many bits are removed.

Original

Vec::drain

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the bit-vector.

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 1];
let bv2: BitVec = bv.drain(1 ..).collect();
assert_eq!(bv, bits![0]);
assert_eq!(bv2, bits![1, 1]);

// A full range clears the vector
bv.drain(..);
assert_eq!(bv, bits![]);

pub fn clear(&mut self)

Clears the bit-vector, removing all values.

Note that this method has no effect on the allocated capacity of the bit-vector.

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 1];
bv.clear();
assert!(bv.is_empty());

pub fn len(&self) -> usize

Returns the number of bits in the bit-vector, also referred to as its ‘length’.

Original

Vec::len

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 0, 1];
assert_eq!(bv.len(), 3);

pub fn is_empty(&self) -> bool

Returns true if the bit-vector contains no bits.

Original

Vec::is_empty

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();
assert!(bv.is_empty());

bv.push(true);
assert!(!bv.is_empty());

pub fn split_off(&mut self, at: usize) -> Self

Splits the collection into two at the given index.

Returns a newly allocated bit-vector containing the bits in range [at, len). After the call, the original bit-vector will be left containing the bits [0, at) with its previous capacity unchanged.

Original

Vec::split_off

Panics

Panics if at > len.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1];
let bv2 = bv.split_off(1);
assert_eq!(bv, bits![0]);
assert_eq!(bv2, bits![0, 1]);

pub fn resize_with<F>(&mut self, new_len: usize, func: F) where
F: FnMut() -> bool,

Resizes the BitVec in-place so that len is equal to new_len.

If new_len is greater than len, the BitVec is extended by the difference, with each additional slot filled with the result of calling the closure func. The return values from func will end up in the BitVec in the order they have been generated.

If new_len is less than len, the BitVec is simply truncated.

This method uses a closure to create new values on every push. If you’d rather Clone a given value, use resize. If you want to use the Default trait to generate values, you can pass Default::default() as the second argument.

Original

Vec::resize_with

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 3];
bv.resize_with(5, Default::default);
assert_eq!(bv, bits![1, 1, 1, 0, 0]);

let mut bv = bitvec![];
let mut p = 0;
bv.resize_with(4, || { p += 1; p % 2 == 0 });
assert_eq!(bv, bits![0, 1, 0, 1]);

pub fn leak<'a>(self) -> &'a mut BitSlice<O, T>

Consumes and leaks the BitVec, returning a mutable reference to the contents, &'a mut BitSlice<O, T>. This lifetime may be chosen to be 'static.

This function is similar to the leak function on BitBox.

This function is mainly useful for data that lives for the remainder of the program’s life. Dropping the returned reference will cause a memory leak.

Original

Vec::leak

Examples

Simple usage:

use bitvec::prelude::*;

let x = bitvec![0, 0, 1];
let static_ref: &'static mut BitSlice = x.leak();
static_ref.set(0, true);
assert_eq!(static_ref, bits![1, 0, 1]);

pub fn resize(&mut self, new_len: usize, value: bool)

Resizes the BitVec in-place so that len is equal to new_len.

If new_len is greater than len, the BitVec is extended by the difference, with each additional slot filled with value. If new_len is less than len, the BitVec is simply truncated.

This method requires a single bool value. If you need more flexibility, use resize_with.

Original

Vec::resize

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1];
bv.resize(3, false);
assert_eq!(bv, bits![1, 0, 0]);

let mut bv = bitvec![1; 4];
bv.resize(2, false);
assert_eq!(bv, bits![1; 2]);

pub fn resize_default(&mut self, new_len: usize)

👎 Deprecated:
Vec::resize_default is deprecated

Resizes the BitVec in-place so that len is equal to new_len.

pub fn splice<R, I>(
&mut self,
range: R,
replace_with: I
) -> Splice<'_, O, T, I::IntoIter> where
R: RangeBounds<usize>,
I: IntoIterator<Item = bool>,

Creates a splicing iterator that replaces the specified range in the bit-vector with the given replace_with iterator and yields the removed items. replace_with does not need to be the same length as range.

range is removed even if the iterator is not consumed until the end.

It is unspecified how many bits are removed from the vector if the Splice value is leaked.

The input iterator replace_with is only consumed when the Splice value is dropped.

This is optimal if:

the tail (bits in the vector after range) is empty

or replace_with yields fewer bits than range’s length

or the lower bound of its size_hint is exact.

Otherwise, a temporary bit-vector is allocated and the tail is moved twice.

Original

Vec::splice

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the bit-vector.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0];
let new = bits![1, 0];
let old: BitVec = bv.splice(.. 2, new.iter().by_val()).collect();
assert_eq!(bv, bits![1, 0, 0]);
assert_eq!(old, bits![0, 1]);

impl<O, T> BitVec<O, T> where
O: BitOrder,
T: BitStore,

General-purpose functions not present on Vec<T>.

pub fn repeat(bit: bool, len: usize) -> Self

Constructs a BitVec from a value repeated many times.

This function is equivalent to the bitvec![O, T; bit; len] macro call, and is in fact the implementation of that macro syntax.

Parameters

bit: The bit value to which all len allocated bits will be set.

len: The number of live bits in the constructed BitVec.

Returns

A BitVec with len live bits, all set to bit.

Examples

use bitvec::prelude::*;

let bv = BitVec::<Msb0, u8>::repeat(true, 20);
assert_eq!(bv, bits![1; 20]);

pub fn from_bitslice(slice: &BitSlice<O, T>) -> Self

Copies the contents of a BitSlice into a new allocation.

This is an exact copy: the newly-created bit-vector is initialized with a direct copy of the slice’s underlying contents, and its handle is set to use slice’s head index. Slices that do not begin at the zeroth bit of the base element will thus create misaligned vectors.

You can move the bit-vector contents down to begin at the zero index of the bit-vector’s buffer with force_align.

Examples

use bitvec::prelude::*;

let bits = bits![0, 1, 0, 1, 1, 0, 1, 1];
let bv = BitVec::from_bitslice(&bits[2 ..]);
assert_eq!(bv, bits[2 ..]);
assert_eq!(bits.as_raw_slice(), bv.as_raw_slice());

pub fn from_element(elem: T) -> Self

Constructs a new BitVec from the bit-pattern of a single element.

This function copies elem into a new vector, then views that vector as bits.

Examples

use bitvec::prelude::*;

assert_eq!(BitVec::<Msb0, _>::from_element(0xABBAu16).count_ones(), 10);

pub fn from_slice(slice: &[T]) -> Result<Self, BitSpanError<T>>

Constructs a new BitVec from the bit-pattern of an element slice.

This function copies slice into a new vector, then views that vector as bits.

Parameters

slice: A slice of elements. It should not exceed BitSlice::<O, T>::MAX_ELTS.

Returns

This returns an error if BitSlice::<O, T>::from_slice fails; otherwise, it returns the newly allocated and initialized bit-vector.

Examples

use bitvec::prelude::*;

let slice = &[0u8, 1, 2, 3];
let bv = BitVec::<Lsb0, _>::from_slice(slice);
assert!(bv.is_ok());
assert_eq!(bv.unwrap().len(), 32);

pub fn from_vec(vec: Vec<T>) -> Self

Converts a Vec<T> into a BitVec<O, T> without copying its buffer.

Parameters

vec: A vector to view as bits.

Returns

A BitVec over the vec buffer.

Panics

This panics if vec is too long to convert into a BitVec. See BitSlice::MAX_ELTS.

Examples

use bitvec::prelude::*;

let vec = vec![0u8; 4];
let bv = BitVec::<LocalBits, _>::from_vec(vec);
assert_eq!(bv, bits![0; 32]);

pub fn try_from_vec(vec: Vec<T>) -> Result<Self, Vec<T>>

Converts a Vec<T> into a BitVec<O, T> without copying its buffer.

This method takes ownership of a memory buffer and enables it to be used as a bit-vector. Because Vec can be longer than BitVecs, this is a fallible method, and the original vector will be returned if it cannot be converted.

Parameters

vec: Some vector of memory, to be viewed as bits.

Returns

If vec is short enough to be viewed as a BitVec, then this returns a BitVec over the vec buffer. If vec is too long, then this returns vec unmodified.

Examples

use bitvec::prelude::*;

let vec = vec![0u8; 4];
let bv = BitVec::<LocalBits, _>::try_from_vec(vec).unwrap();
assert_eq!(bv, bits![0; 32]);

An example showing this function failing would require an allocation exceeding !0usize >> 3 bytes in size, which is infeasible to produce.

pub fn extend_from_bitslice<O2, T2>(&mut self, other: &BitSlice<O2, T2>) where
O2: BitOrder,
T2: BitStore,

Copies all bits in a BitSlice into the BitVec.

Original

Vec::extend_from_slice

Type Parameters

This can extend from a BitSlice of any type arguments. Where the source &BitSlice matches self’s type parameters, the implementation is able to attempt to accelerate the copy; however, if the type parameters do not match, then the implementation falls back to a bit-by-bit iteration and is equivalent to the Extend implementation.

You should only use this method when the type parameters match and there is a possibility of copy acceleration. Otherwise, .extend() is the correct API.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1];
bv.extend_from_bitslice(bits![1, 1, 0, 1]);

assert_eq!(bv, bits![0, 1, 1, 1, 0, 1]);

pub fn extend_from_raw_slice(&mut self, slice: &[T])

Appends a slice of elements T to the BitVec.

The slice is interpreted as a BitSlice<O, T>, then appended directly to the bit-vector.

Original

Vec::extend_from_slice

pub fn elements(&self) -> usize

Gets the number of elements T that contain live bits of the bit-vector.

Examples

use bitvec::prelude::*;

let bv = bitvec![LocalBits, u16; 1; 50];
assert_eq!(bv.elements(), 4);

pub fn into_boxed_bitslice(self) -> BitBox<O, T>

Converts the bit-vector into BitBox<O, T>.

Note that this will drop any excess capacity.

Original

Vec::into_boxed_slice

API Differences

This returns a bitvec boxed bit-slice, not a standard boxed slice. To convert the underlying buffer into a boxed element slice, use .into_boxed_bitslice().into_boxed_slice().

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0, 0, 1];
let bitslice = bv.into_boxed_slice();

Any excess capacity is removed:

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(100);
bv.extend([0, 1, 0, 0, 1].iter().copied());

assert!(bv.capacity() >= 100);
let bs = bv.into_boxed_bitslice();
assert!(bs.into_bitvec().capacity() >= 5);

pub fn into_vec(self) -> Vec<T>

Removes the bit-precision view, returning the underlying Vec.

Examples

use bitvec::prelude::*;

let bv = bitvec![Lsb0, u8; 0, 1, 0, 0, 1];
let vec = bv.into_vec();
assert_eq!(vec, &[18]);

pub fn set_elements(&mut self, element: T::Mem)

Writes a value into every element that the bit-vector considers live.

This unconditionally writes element into each live location in the backing buffer, without altering the BitVec’s length or capacity.

It is unspecified what effects this has on the allocated but dead elements in the buffer. You may not rely on them being zeroed or being set to the value integer.

Parameters

&mut self

element: The value which will be written to each live location in the bit-vector’s buffer.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![LocalBits, u8; 0; 10];
assert_eq!(bv.as_raw_slice(), [0, 0]);
bv.set_elements(0xA5);
assert_eq!(bv.as_raw_slice(), [0xA5, 0xA5]);

pub fn set_uninitialized(&mut self, value: bool)

Sets the uninitialized bits of the bit-vector to a fixed value.

This method modifies all bits in the allocated buffer that are outside the as_bitslice view so that they have a consistent value. This can be used to zero the uninitialized memory so that when viewed as a raw memory slice, bits outside the live region have a predictable value.

Examples

use bitvec::prelude::*;

let mut bv = 220u8.view_bits::<Lsb0>().to_bitvec();
assert_eq!(bv.as_raw_slice(), &[220u8]);

bv.truncate(4);
assert_eq!(bv.count_ones(), 2);
assert_eq!(bv.as_raw_slice(), &[220u8]);

bv.set_uninitialized(false);
assert_eq!(bv.as_raw_slice(), &[12u8]);

bv.set_uninitialized(true);
assert_eq!(bv.as_raw_slice(), &[!3u8]);

pub fn force_align(&mut self)

Ensures that the live region of the bit-vector’s contents begins at the leading edge of the buffer.

Examples

use bitvec::prelude::*;

let data = 0x3Cu8;
let bits = data.view_bits::<Msb0>();

let mut bv = bits[2 .. 6].to_bitvec();
assert_eq!(bv, bits[2 .. 6]);
assert_eq!(bv.as_raw_slice()[0], data);

bv.force_align();
assert_eq!(bv, bits[2 .. 6]);
// It is not specified what happens
// to bits that are no longer used.
assert_eq!(bv.as_raw_slice()[0] & 0xF0, 0xF0);

pub fn as_bitslice(&self) -> &BitSlice<O, T>

Extracts a bit-slice containing the entire bit-vector.

Equivalent to &bv[..].

Original

Vec::as_slice

API Differences

This returns a bitvec bit-slice, not a standard slice. To view the underlying element buffer, use as_raw_slice.

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0, 0, 1];
let bits = bv.as_bitslice();

pub fn as_mut_bitslice(&mut self) -> &mut BitSlice<O, T>

Extracts a mutable bit-slice of the entire bit-vector.

Equivalent to &mut bv[..].

Original

Vec::as_mut_slice

API Differences

This returns a bitvec bit-slice, not a standard slice. To view the underlying element buffer, use as_mut_raw_slice.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
let bits = bv.as_mut_bitslice();

pub fn as_bitptr(&self) -> BitPtr<Const, O, T>

Returns a raw pointer to the bit-vector’s buffer.

The caller must ensure that the bit-vector outlives the bit-pointer this function returns, or else it will end up pointing to garbage. Modifying the bit-vector may cause its buffer to be reällocated, which would also make any bit-pointers to it invalid.

The caller must also ensure that the memory the bit-pointer (non-transitively) points to is never written to (except inside an UnsafeCell) using this bit-pointer or any bit-pointer derived from it. If you need to mutate the contents of the buffer, use as_mut_bitptr.

Original

Vec::as_ptr

API Differences

This returns a bitvec bit-pointer, not a standard pointer. To take the address of the underlying element buffer, use as_raw_ptr.

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0, 0, 1];
let bp = bv.as_bitptr();

unsafe {
  for i in 0 .. bv.len() {
    assert_eq!(bp.add(i).read(), bv[i]);
  }
}

pub fn as_mut_bitptr(&mut self) -> BitPtr<Mut, O, T>

Returns an unsafe mutable bit-pointer to the bit-vector’s region.

The caller must ensure that the bit-vector outlives the bit-pointer this function returns, or else it will end up pointing to garbage. Modifying the bit-vector may cause its buffer to be reällocated, which would also make any bit-pointers to it invalid.

Original

Vec::as_mut_ptr

API Differences

This returns a bitvec bit-pointer, not a standard pointer. To take the address of the underlying element buffer, use as_mut_raw_ptr.

Examples

use bitvec::prelude::*;

let mut bv = BitVec::<Msb0, u8>::with_capacity(4);
let bp = bv.as_mut_bitptr();
unsafe {
  for i in 0 .. 4 {
    bp.add(i).write(true);
  }
  bv.set_len(4);
}
assert_eq!(bv, bits![1; 4]);

pub fn as_raw_slice(&self) -> &[T]

Views the underlying buffer as a shared element slice.

Original

Vec::as_slice

API Differences

This method is renamed in order to emphasize the semantic distinction between borrowing the bit-vector contents, and borrowing the memory that implements the collection contents.

Examples

use bitvec::prelude::*;

let bv = bitvec![Msb0, u8; 0, 1, 0, 0, 1, 1, 0, 1];
let raw = bv.as_raw_slice();
assert_eq!(raw, &[0x4D]);

pub fn as_mut_raw_slice(&mut self) -> &mut [T]

Views the underlying buffer as an exclusive element slice.

Original

Vec::as_mut_slice

API Differences

This method is renamed in order to emphasize the semantic distinction between borrowing the bit-vector contents, and borrowing the memory that implements the collection contents.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![Msb0, u8; 0, 1, 0, 0, 1, 1, 0, 1];
let raw = bv.as_mut_raw_slice();
assert_eq!(raw, &[0x4D]);
raw[0] = 0xD4;
assert_eq!(bv, bits![1, 1, 0, 1, 0, 1, 0, 0]);

pub fn as_raw_ptr(&self) -> *const T

Returns a raw pointer to the bit-vector’s buffer.

Original

Vec::as_ptr

API Differences

This method is renamed in order to emphasize the semantic distinction between taking a pointer to the start of the bit-vector contents, and taking a pointer to the underlying memory that implements the collection contents.

Examples

use bitvec::prelude::*;

let bv = bitvec![Msb0, u8; 0, 1, 0, 0, 1];
let addr = bv.as_raw_ptr();

pub fn as_mut_raw_ptr(&mut self) -> *mut T

Returns an unsafe mutable pointer to the bit-vector’s buffer.

Original

Vec::as_mut_ptr

API Differences

This method is renamed in order to emphasize the semantic distinction between taking a pointer to the start of the bit-vector contents, and taking a pointer to the underlying memory that implements the collection contents.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 0, 1];
let addr = bv.as_mut_raw_ptr();