Struct bitvec::vec::BitVec

#[repr(C)]
pub struct BitVec<C = BigEndian, T = u8> where
    C: Cursor,
    T: Bits,  { /* fields omitted */ }

A compact Vec of bits, whose cursor and storage type can be customized.

BitVec is a newtype wrapper over Vec, and as such is exactly three words in size on the stack.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();
bv.push(false);
bv.push(true);

assert_eq!(bv.len(), 2);
assert_eq!(bv[0], false);

assert_eq!(bv.pop(), Some(true));
assert_eq!(bv.len(), 1);

bv.set(0, true);
assert_eq!(bv[0], true);

bv.extend([0u8, 1, 0].iter().map(|n| *n != 0u8));
for bit in &*bv {
  println!("{}", bit);
}
assert_eq!(bv, bitvec![1, 0, 1, 0]);

The bitvec! macro is provided to make initialization more convenient.

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 2, 3];
bv.push(false);
assert_eq!(bv, bitvec![0, 1, 1, 1, 0]);

It can also initialize each element of a BitVec<_, T> with a given value. This may be more efficient than performing allocation and initialization in separate steps, especially when initializing a vector of zeros:

use bitvec::prelude::*;

let bv = bitvec![0; 15];
assert_eq!(bv, bitvec![0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]);

// The following is equivalent, but potentially slower:
let mut bv1: BitVec = BitVec::with_capacity(15);
bv1.resize(15, false);

Use a BitVec<T> as an efficient stack:

use bitvec::prelude::*;
let mut stack: BitVec = BitVec::new();

stack.push(false);
stack.push(true);
stack.push(true);

while let Some(top) = stack.pop() {
  //  Prints true, true, false
  println!("{}", top);
}

Indexing

The BitVec type allows you to access values by index, because it implements the Index trait. An example will be more explicit:

use bitvec::prelude::*;

let bv = bitvec![0, 0, 1, 1];
println!("{}", bv[1]); // it will display 'false'

However, be careful: if you try to access an index which isn’t in the BitVec, your software will panic! You cannot do this:

use bitvec::prelude::*;

let bv = bitvec![0, 1, 0, 1];
println!("{}", bv[6]); // it will panic!

In conclusion: always check if the index you want to get really exists before doing it.

Slicing

A BitVec is growable. A BitSlice, on the other hand, is fixed size. To get a bit slice, use &. Example:

use bitvec::prelude::*;
fn read_bitslice(slice: &BitSlice) {
	// use slice
}

let bv = bitvec![0, 1];
read_bitslice(&bv);

// … and that’s all!
// you can also do it like this:
let bs : &BitSlice = &bv;

In Rust, it’s more common to pass slices as arguments rather than vectors when you do not want to grow or shrink it. The same goes for Vec and [&[]], and String and &str.

Capacity and Reallocation

The capacity of a bit vector is the amount of space allocated for any future bits that will be added onto the vector. This is not to be confused with the length of a vector, which specifies the number of live, useful bits within the vector. If a vector’s length exceeds its capacity, its capacity will automatically be increased, but its storage elements will have to be reallocated.

For example, a bit vector with capacity 10 and length 0 would be an allocated, but uninhabited, vector, with space for ten more bits. Pushing ten or fewer bits onto the vector will not change its capacity or cause reallocation to occur. However, if the vector’s length is increased to eleven, it will have to reallocate, which can be slow. For this reason, it is recommended to use BitVec::with_capacity whenever possible to specify how big the bit vector is expected to get.

Guarantees

Due to its incredibly fundamental nature, BitVec makes a lot of guarantees about its design. This ensures that it is as low-overhead as possible in the general case, and can be correctly manipulated in fundamental ways by unsafe code.

Most fundamentally, BitVec is and always will be a ([BitPtr], capacity) doublet. No more, no less. The order of these fields is unspecified, and you should only interact with the members through the provided APIs. Note that BitPtr is not directly manipulable, and must never be written or interpreted as anything but opaque binary data by user code.

When a BitVec has allocated memory, then the memory to which it points is on the heap (as defined by the allocator Rust is configured to use by default), and its pointer points to len initialized bits in order of the Cursor type parameter, followed by capacity - len logically uninitialized bits.

BitVec will never perform a “small optimization” where elements are stored in its handle representation, for two reasons:

It would make it more difficult for user code to correctly manipulate a BitVec. The contents of the BitVec would not have a stable address if the handle were moved, and it would be more difficult to determine if a BitVec had allocated memory.

It would penalize the general, heap-allocated, case by incurring a branch on every access.

BitVec will never automatically shrink itself, even if it is emptied. This ensures that no unnecessary allocations or deallocations occur. Emptying a BitVec and then refilling it to the same length will incur no calls to the allocator. If you wish to free up unused memory, use shrink_to_fit.

Erasure

BitVec will not specifically overwrite any data that is removed from it, nor will it specifically preserve it. Its uninitialized memory is scratch space that may be used however the implementation desires, and must not be relied upon as stable. Do not rely on removed data to be erased for security purposes. Even if you drop a BitVec, its buffer may simply be reused for other data structures in your program. Even if you zero a BitVec’s memory first, that may not actually occur if the optimizer does not consider this an observable side effect. There is one case that will never break, however: using unsafe to construct a [T] slice over the BitVec’s capacity, and writing to the excess space, then increasing the length to match, is always valid.

Type Parameters

C: Cursor: An implementor of the Cursor trait. This type is used to convert semantic indices into concrete bit positions in elements, and store or retrieve bit values from the storage type.

T: Bits: An implementor of the Bits trait: u8, u16, u32, u64. This is the actual type in memory the slice will use to store data.

Safety

The BitVec handle has the same size as standard Rust Vec handles, but it is extremely binary incompatible with them. Attempting to treat BitVec<_, T> as Vec<T> in any manner except through the provided APIs is catastrophically unsafe and unsound.

[&[]]: https://doc.rust-lang.org/stable/std/primitive.slice.html

Methods

impl<C, T> BitVec<C, T> where
C: Cursor,
T: Bits,

pub fn new() -> Self

Constructs a new, empty, BitVec<C, T>.

The vector does not allocate until bits are written into it.

Returns

An empty, unallocated, BitVec handle.

Examples

use bitvec::prelude::*;

let bv: BitVec = BitVec::new();
assert!(bv.is_empty());
assert_eq!(bv.capacity(), 0);

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty, BitVec<T> with the specified capacity.

The new vector will be able to hold at least capacity elements before it reallocates. If capacity is 0, it will not allocate.

Parameters

capacity: The minimum number of bits that the new vector will need to be able to hold.

Returns

An empty vector with at least the given capacity.

Examples

use bitvec::prelude::*;

let bv: BitVec = BitVec::with_capacity(10);
assert!(bv.is_empty());
assert!(bv.capacity() >= 10);

pub fn from_element(elt: T) -> Self

Constructs a BitVec from a single element.

The produced BitVec will span the element, and include all bits in it.

Parameters

elt: The source element.

Returns

A BitVec over the provided element.

Examples

use bitvec::prelude::*;

let bv = BitVec::<BigEndian, u8>::from_element(5);
assert_eq!(bv.count_ones(), 2);

pub fn from_slice(slice: &[T]) -> Self

Constructs a BitVec from a slice of elements.

The produced BitVec will span the provided slice.

Parameters

slice: The source elements to copy into the new BitVec.

Returns

A BitVec set to the provided slice values.

Examples

use bitvec::prelude::*;

let src: &[u8] = &[5, 10];
let bv: BitVec = src.into();
assert!(bv[5]);
assert!(bv[7]);
assert!(bv[12]);
assert!(bv[14]);

pub fn from_vec(vec: Vec<T>) -> Self

Consumes a Vec<T> and creates a BitVec<C, T> from it.

Parameters

vec: The source vector whose memory will be used.

Returns

A new BitVec using the vec Vec’s memory.

Panics

Panics if the source vector would cause the BitVec to overflow capacity.

Examples

use bitvec::prelude::*;

let src: Vec<u8> = vec![1, 2, 4, 8];
let bv: BitVec = src.into();
assert_eq!(
  "[00000001, 00000010, 00000100, 00001000]",
  &format!("{}", bv),
);

pub fn from_bitslice(slice: &BitSlice<C, T>) -> Self

Clones a &BitSlice into a BitVec.

Parameters

slice

Returns

A BitVec containing the same bits as the source slice.

Examples

use bitvec::prelude::*;

let src: &[u8] = &[0, !0];
let bs: &BitSlice = src.into();
let bv = BitVec::from_bitslice(bs);
assert_eq!(bv.len(), 16);
assert!(bv.some());

pub fn from_boxed_bitslice(slice: BitBox<C, T>) -> Self

Converts a frozen BitBox allocation into a growable BitVec.

This does not copy or reallocate.

Parameters

slice: A BitBox to be thawed.

Returns

A growable collection over the original memory of the slice.

pub unsafe fn from_raw_parts(pointer: BitPtr<T>, capacity: usize) -> Self

Creates a new BitVec<C, T> directly from the raw parts of another.

Parameters

pointer: The BitPtr<T> to use.

capacity: The number of T elements allocated in that slab.

Returns

A BitVec over the given slab of memory.

Safety

This is highly unsafe, due to the number of invariants that aren’t checked:

pointer needs to have been previously allocated by some allocating type.

pointer’s T needs to have the same size and alignment as it was initially allocated.

pointer’s element count needs to be less than or equal to the original allocation capacity.

capacity needs to be the original allocation capacity for the pointer.

Violating these will cause problems, like corrupting the handle’s concept of memory, the allocator’s internal data structures, and the sanity of your program. It is absolutely not safe to construct a BitVec whose T differs from the type used for the initial allocation.

The ownership of pointer is effectively transferred to the BitVec<C, T> which may then deallocate, reallocate, or modify the contents of the referent slice at will. Ensure that nothing else uses the pointer after calling this function.

pub fn capacity(&self) -> usize

Returns the number of bits the vector can hold without reallocating.

Parameters

&self

Returns

The number of bits that the vector can hold before reallocating.

Examples

use bitvec::prelude::*;

let bv: BitVec = BitVec::with_capacity(10);
assert!(bv.is_empty());
assert!(bv.capacity() >= 10);

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more bits to be inserted.

The collection may reserve more space to avoid frequent reallocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Parameters

&mut self

additional: The number of extra bits to be granted space.

Panics

Panics if the new capacity would overflow the vector’s limits.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 5];
assert!(bv.capacity() >= 5);
bv.reserve(10);
assert!(bv.capacity() >= 15);

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for at least additional more bits.

After calling reserve_exact, the capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore, the capacity cannot be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Parameters

&mut self

additional: The number of extra bits to be granted space.

Panics

Panics if the new capacity would overflow the vector’s limits.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 5];
assert!(bv.capacity() >= 5);
bv.reserve_exact(10);
assert!(bv.capacity() >= 15);

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the vector as much as possible.

It will drop down as close as possible to the length, but the allocator may still inform the vector that there is space for bits.

This does not modify the contents of the memory store! It will not zero any memory that had been used and then removed from the vector’s live count.

Parameters

&mut self

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 100];
let cap = bv.capacity();
bv.truncate(10);
bv.shrink_to_fit();
assert!(bv.capacity() <= cap);

pub fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len bits and dropping the rest.

If len is greater than the vector’s current length, this has no effect.

Parameters

&mut self

len: The new length of the vector.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 15];
bv.truncate(10);
assert_eq!(bv.len(), 10);

bv.truncate(15);
assert_eq!(bv.len(), 10);

pub fn as_bitslice(&self) -> &BitSlice<C, T>

Produces a BitSlice containing the entire vector.

Equivalent to &s[..].

Parameters

&self

Returns

A BitSlice over the vector.

Examples

use bitvec::prelude::*;

let bv = bitvec![0, 1, 1, 0];
let bs = bv.as_bitslice();

pub fn as_mut_bitslice(&mut self) -> &mut BitSlice<C, T>

Produces a mutable BitSlice containing the entire vector.

Equivalent to &mut s[..].

Parameters

&mut self

Returns

A mutable BitSlice over the vector.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 1, 0];
let bs = bv.as_mut_bitslice();

pub unsafe fn set_len(&mut self, len: usize)

Sets the length of the vector.

This unconditionally sets the size of the vector, without modifying its contents. It is up to the caller to ensure that the vector’s buffer can hold the new size.

Parameters

&mut self

len: The new length of the vector. This must be less than the maximum number of bits that the vector can hold.

Panics

This panics if len overflows the vector's intrinsic or allocated capacities.

Safety

The caller must ensure that the new length is sound for the vector.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::with_capacity(15);
assert!(bv.is_empty());
unsafe { bv.set_len(10) };
assert_eq!(bv.len(), 10);

pub fn swap_remove(&mut self, index: usize) -> bool

Removes a bit from the vector and returns it.

The removed bit is replaced by the last bit in the vector.

Parameters

&mut self

index: The index whose bit is to be returned, and replaced by the tail.

Returns

The bit at the requested index.

Panics

Panics if the index is out of bounds.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 0, 0, 1];
assert!(!bv[2]);
assert_eq!(bv.len(), 5);
assert!(!bv.swap_remove(2));
assert!(bv[2]);
assert_eq!(bv.len(), 4);

pub fn insert(&mut self, index: usize, value: bool)

Inserts an element at a position, shifting all elements after it to the right.

Note that this is O(n) runtime.

Parameters

&mut self

index: The position at which to insert. This may be any value from 0 up to and including self.len(). At self.len(), it is equivalent to calling self.push(value).

value: The value to be inserted.

Panics

Panics if index is greater than the length.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 0, 0];
bv.insert(2, true);
assert_eq!(bv, bitvec![0, 0, 1, 0, 0]);
bv.insert(5, true);
assert_eq!(bv, bitvec![0, 0, 1, 0, 0, 1]);

pub fn remove(&mut self, index: usize) -> bool

Removes and returns the element at position index, shifting all elements after it to the left.

Parameters

&mut self

index: The position whose bit is to be removed. This must be in the domain 0 .. self.len().

Returns

The bit at the requested index.

Panics

Panics if index is out of bounds for the vector.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1, 0, 0];
assert!(bv.remove(2));
assert_eq!(bv, bitvec![0, 0, 0, 0]);

pub fn retain<F>(&mut self, pred: F) where
F: FnMut(bool) -> bool,

Retains only the bits that pass the predicate.

This removes all bits b where f(e) returns false. This method operates in place and preserves the order of the retained bits. Because it is in-place, it operates in O(n²) time.

Parameters

&mut self

pred: The testing predicate for each bit.

Type Parameters

F: FnMut(bool) -> bool: A function that can be invoked on each bit, returning whether the bit should be kept or not.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 1, 0, 1, 0, 1];
bv.retain(|b| b);
assert_eq!(bv, bitvec![1, 1, 1]);

pub fn push(&mut self, value: bool)

Appends a bit to the back of the vector.

If the vector is at capacity, this may cause a reallocation.

Parameters

&mut self

value: The bit value to append.

Panics

This will panic if the push will cause the vector to allocate above BitPtr<T> or machine capacity.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();
assert!(bv.is_empty());
bv.push(true);
assert_eq!(bv.len(), 1);
assert!(bv[0]);

pub fn pop(&mut self) -> Option<bool>

Removes the last bit from the collection, if present.

Parameters

&mut self

Returns

If the vector is not empty, this returns the last bit; if it is empty, this returns None.

Examples

use bitvec::prelude::*;

let mut bv: BitVec = BitVec::new();
assert!(bv.is_empty());
bv.push(true);
assert_eq!(bv.len(), 1);
assert!(bv[0]);

assert!(bv.pop().unwrap());
assert!(bv.is_empty());
assert!(bv.pop().is_none());

pub fn append<D, U>(&mut self, other: &mut BitVec<D, U>) where
D: Cursor,
U: Bits,

Moves all the elements of other into self, leaving other empty.

Parameters

&mut self

other: A BitVec of any order and storage type. Its bits are appended to self.

Panics

Panics if the joined vector is too large.

Examples

use bitvec::prelude::*;

let mut bv1 = bitvec![0; 10];
let mut bv2 = bitvec![1; 10];
bv1.append(&mut bv2);
assert_eq!(bv1.len(), 20);
assert!(bv1[10]);
assert!(bv2.is_empty());

Important traits for Drain<'a, C, T>
impl<'a, C, T> Iterator for Drain<'a, C, T> where
C: Cursor,
T: 'a + Bits, type Item = bool;

pub fn drain<R>(&mut self, range: R) -> Drain<C, T> where
R: RangeBounds<usize>,

Creates a draining iterator that removes the specified range from the vector and yields the removed bits.

Notes

The element range is removed, regardless of whether the iterator is consumed.

The amount of items removed from the vector if the draining iterator is leaked, is left unspecified.

Parameters

&mut self

range: any range literal, which is used to define the range of the vector that is drained.

Returns

An iterator over the specified range.

Panics

Panics if the range is ill-formed, or if it is beyond the vector bounds.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0, 0, 1, 1, 1, 0, 0];
assert_eq!(bv.len(), 7);
for bit in bv.drain(2 .. 5) {
  assert!(bit);
}
assert!(bv.not_any());
assert_eq!(bv.len(), 4);

pub fn clear(&mut self)

Clears the vector, removing all values.

Note that this method has no effect on the allocated capacity of the vector.

Parameters

&mut self

Effects

Becomes an uninhabited slice.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![1; 30];
assert_eq!(bv.len(), 30);
assert!(bv.iter().all(|b| b));
bv.clear();
assert!(bv.is_empty());

After calling clear(), bv will no longer show raw memory, so the above test cannot show that the underlying memory is not altered. This is also an implementation detail on which you should not rely.

pub fn split_off(&mut self, at: usize) -> Self

Splits the collection into two at the given index.

Returns a newly allocated Self. self contains elements [0, at), and the returned Self contains elements [at, self.len()).

Note that the capacity of self does not change.

Parameters

&mut self

at: The index at which to perform the split. This must be in the domain 0 ..= self.len(). When it is self.len(), an empty vector is returned.

Returns

A new BitVec containing all the elements from at onwards.

Panics

Panics if at is beyond self.len().

Examples

use bitvec::prelude::*;

let mut bv1 = bitvec![0, 0, 0, 1, 1, 1];
let bv2 = bv1.split_off(3);
assert_eq!(bv1, bitvec![0, 0, 0]);
assert_eq!(bv2, bitvec![1, 1, 1]);

pub fn resize(&mut self, new_len: usize, value: bool)

Resizes the BitVec in place so that len is equal to new_len.

If new_len is greater than len, then the vector is extended by the difference, and filled with the provided value. If new_len is less than len, then the vector is just truncated.

Parameters

&mut self

new_len: The new length of the vector.

value: The fill value if the vector is to be extended.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0; 4];
bv.resize(8, true);
assert_eq!(bv, bitvec![0, 0, 0, 0, 1, 1, 1, 1]);
bv.resize(5, false);
assert_eq!(bv, bitvec![0, 0, 0, 0, 1]);

Important traits for Splice<'a, C, T, I>
impl<'a, C, T, I> Iterator for Splice<'a, C, T, I> where
C: Cursor,
T: 'a + Bits,
I: Iterator<Item = bool>, type Item = bool;

pub fn splice<R, I>(
&mut self,
range: R,
replacement: I
) -> Splice<C, T, <I as IntoIterator>::IntoIter> where
R: RangeBounds<usize>,
I: IntoIterator<Item = bool>,

Creates a splicing iterator that exchanges the specified range for the replacement iterator, yielding the removed items. The range and its replacement do not need to be the same size.

Notes

The element range is removed and replaced even if the iterator produced by this method is not fully consumed.

It is unspecified how many bits are removed from the BitVec if the returned iterator is leaked.

The input iterator replacement is only consumed when the returned iterator is dropped.

This is optimal if:

the tail (elements in the BitVec after range) is empty,

replace_with yields fewer characters than range’s length,

the lower bound of replacement.size_hint() is exact.

Parameters

&mut self

range: A range of indices in the BitVec to pull out of the collection.

replacement: Something which can be used to provide new bits to replace the removed range.

The entirety of replacement will be inserted into the slot marked by range. If replacement is an infinite iterator, then this will hang, and crash your program.

Returns

An iterator over the bits marked by range.

Panics

Panics if the range is ill-formed, or extends past the end of the BitVec.

Examples

This example starts with six bits of zero, and then splices out bits 2 and 3 and replaces them with four bits of one.

use bitvec::prelude::*;

let mut bv = bitvec![0; 6];
let bv2 = bitvec![1; 4];

let s = bv.splice(2 .. 4, bv2).collect::<BitVec>();
assert_eq!(s.len(), 2);
assert!(!s[0]);
assert_eq!(bv, bitvec![0, 0, 1, 1, 1, 1, 0, 0]);

pub fn set_elements(&mut self, element: T)

Sets the backing storage to the provided element.

This unconditionally sets each allocated element in the backing storage to the provided value, without altering the BitVec length or capacity. It operates on the underlying Vec’s memory region directly, and will ignore the BitVec’s cursors.

This has the unobservable effect of setting the allocated, but dead, bits beyond the end of the vector’s length, up to its capacity.

Parameters

&mut self

element: The value to which each allocated element in the backing store will be set.

Examples

use bitvec::prelude::*;

let mut bv = bitvec![0; 10];
assert_eq!(bv.as_slice(), &[0, 0]);
bv.set_elements(0xA5);
assert_eq!(bv.as_slice(), &[0xA5, 0xA5]);

Important traits for BitVec<C, T>
impl<C, T> Write for BitVec<C, T> where
C: Cursor,
T: Bits,

pub fn change_cursor<D>(self) -> BitVec<D, T> where
D: Cursor,

Changes the cursor type on the vector handle, without changing its contents.

Parameters

self

Returns

An equivalent vector handle with a new cursor type. The contents of the backing storage are unchanged.

To reorder the bits in memory, drain this vector into a new handle with the desired cursor type.

pub fn into_boxed_bitslice(self) -> BitBox<C, T>

Degrades a BitVec to a BitBox, freezing its size.

Parameters

self

Returns

Itself, with its size frozen and ungrowable.

pub fn into_boxed_slice(self) -> Box<[T]>

Degrades a BitVec to a standard boxed slice.

Parameters

self

Returns

A boxed slice of the data the BitVec had owned.

pub fn into_vec(self) -> Vec<T>

Degrades a BitVec to a standard Vec.

Parameters

self

Returns

The plain vector underlying the BitVec.

pub fn bitptr(&self) -> BitPtr<T>

Gets the raw BitPtr powering the vector.

Parameters

&self

Returns

The underlying BitPtr for the vector.

Notes

The BitPtr<T> return type is opaque, and not exported by the crate. Users are not able to use it in any way except to construct another BitVec<_, T> from it. It is not possible for user code to even express the name of the type.

use bitvec::prelude::*;
use std::mem;

let bv = bitvec![1; 10];
let bitptr = bv.bitptr();
let cap = bv.capacity();
mem::forget(bv);
let bv2 = unsafe {
  BitVec::<BigEndian, _>::from_raw_parts(bitptr, cap)
};
assert_eq!(bv2.len(), 10);
assert!(bv2[9]);