Trait bitvec::view::BitViewSized

pub trait BitViewSized: BitView + Sized {
    const ELTS: usize;
    const BITS: usize = Self::ELTS * <::Mem as IsNumber>::BITS as usize;
}

Helper for size awareness on Sized storage regions.

Required Associated Constants

const ELTS: usize

Counts the number of elements T contained in the type.

Provided Associated Constants

const BITS: usize = Self::ELTS * <::Mem as IsNumber>::BITS as usize

Counts the number of bits contained in the type.