Trait bitvec::view::BitViewSized

pub trait BitViewSized: BitView + Sized {
    const ZERO: Self;

    // Required methods
    fn as_raw_slice(&self) -> &[Self::Store];
    fn as_raw_mut_slice(&mut self) -> &mut [Self::Store];

    // Provided method
    fn into_bitarray<O>(self) -> BitArray<Self, O>
       where O: BitOrder { ... }
}

Helper trait for scalars and arrays, but not slices.

Required Associated Constants

const ZERO: Self

The zero constant.

Required Methods

fn as_raw_slice(&self) -> &[Self::Store]

Views the type as a slice of its elements.

fn as_raw_mut_slice(&mut self) -> &mut [Self::Store]

Views the type as a mutable slice of its elements.

Provided Methods

fn into_bitarray<O>(self) -> BitArray<Self, O>where O: BitOrder,

Wraps self in a BitArray.

Object Safety

This trait is not object safe.