Trait bitvec::view::BitViewSized

pub trait BitViewSized: BitView + Sized {
    const ZERO: Self;

    fn as_raw_slice(&self) -> &[Self::Store];

    fn as_raw_mut_slice(&mut self) -> &mut [Self::Store];

    fn into_bitarray<O>(self) -> BitArray<Self, O>
    where
        O: BitOrder,
    { ... }
}

Helper trait for scalars and arrays, but not slices.

Associated Constants

const ZERO: Self

The zero constant.

Required methods

fn as_raw_slice(&self) -> &[Self::Store]

Views the type as a slice of its elements.

fn as_raw_mut_slice(&mut self) -> &mut [Self::Store]

Views the type as a mutable slice of its elements.

Provided methods

fn into_bitarray<O>(self) -> BitArray<Self, O> where
O: BitOrder,

Wraps self in a BitArray.